% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm_main.R
\name{AIC,PMM2fit-method}
\alias{AIC,PMM2fit-method}
\title{Calculate AIC for a PMM2fit object}
\usage{
\S4method{AIC}{PMM2fit}(object, ..., k = 2)
}
\arguments{
\item{object}{A PMM2fit object}

\item{...}{Additional arguments (not used)}

\item{k}{The penalty per parameter to be used; default is 2}
}
\value{
The AIC value
}
\description{
Calculate AIC for a PMM2fit object
}
