% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm3_main.R
\name{AIC,PMM3fit-method}
\alias{AIC,PMM3fit-method}
\title{Calculate AIC for PMM3fit object}
\usage{
\S4method{AIC}{PMM3fit}(object, ..., k = 2)
}
\arguments{
\item{object}{PMM3fit object}

\item{...}{Additional arguments (not used)}

\item{k}{Penalty per parameter (default 2)}
}
\value{
AIC value
}
\description{
Calculate AIC for PMM3fit object
}
