% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_ts_methods.R
\name{AIC,TS2fit-method}
\alias{AIC,TS2fit-method}
\title{AIC for TS2fit objects}
\usage{
\S4method{AIC}{TS2fit}(object, ..., k = 2)
}
\arguments{
\item{object}{TS2fit (or subclass) object}

\item{...}{Ignored}

\item{k}{Penalty per parameter (default 2)}
}
\value{
Numeric AIC value
}
\description{
AIC for TS2fit objects
}
