% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm3_ts_methods.R
\name{AIC,TS3fit-method}
\alias{AIC,TS3fit-method}
\title{Calculate AIC for TS3fit object}
\usage{
\S4method{AIC}{TS3fit}(object, ..., k = 2)
}
\arguments{
\item{object}{TS3fit object}

\item{...}{Additional arguments (not used)}

\item{k}{Penalty per parameter (default 2)}
}
\value{
AIC value
}
\description{
Calculate AIC for TS3fit object
}
