% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_main.R
\name{BIC,PMM2fit-method}
\alias{BIC,PMM2fit-method}
\title{Calculate BIC for PMM2fit object}
\usage{
\S4method{BIC}{PMM2fit}(object, ...)
}
\arguments{
\item{object}{PMM2fit object}

\item{...}{Additional arguments (not used)}
}
\value{
BIC value
}
\description{
Calculate BIC for PMM2fit object
}
