% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm3_main.R
\name{BIC,PMM3fit-method}
\alias{BIC,PMM3fit-method}
\title{Calculate BIC for PMM3fit object}
\usage{
\S4method{BIC}{PMM3fit}(object, ...)
}
\arguments{
\item{object}{PMM3fit object}

\item{...}{Additional arguments (not used)}
}
\value{
BIC value
}
\description{
Calculate BIC for PMM3fit object
}
