% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_ts_methods.R
\name{BIC,TS2fit-method}
\alias{BIC,TS2fit-method}
\title{BIC for TS2fit objects}
\usage{
\S4method{BIC}{TS2fit}(object, ...)
}
\arguments{
\item{object}{TS2fit (or subclass) object}

\item{...}{Additional arguments (not used)}
}
\value{
Numeric BIC value
}
\description{
BIC for TS2fit objects
}
