% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm3_ts_methods.R
\name{BIC,TS3fit-method}
\alias{BIC,TS3fit-method}
\title{BIC for TS3fit objects}
\usage{
\S4method{BIC}{TS3fit}(object, ...)
}
\arguments{
\item{object}{TS3fit (or subclass) object}

\item{...}{Additional arguments (not used)}
}
\value{
Numeric BIC value
}
\description{
BIC for TS3fit objects
}
