% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm_main.R
\name{coef,PMM2fit-method}
\alias{coef,PMM2fit-method}
\title{Extract coefficients from a PMM2fit object}
\usage{
\S4method{coef}{PMM2fit}(object, ...)
}
\arguments{
\item{object}{A PMM2fit object}

\item{...}{Additional arguments (not used)}
}
\value{
A vector of coefficients
}
\description{
Extract coefficients from a PMM2fit object
}
