% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm3_main.R
\name{coef,PMM3fit-method}
\alias{coef,PMM3fit-method}
\title{Extract coefficients from PMM3fit object}
\usage{
\S4method{coef}{PMM3fit}(object, ...)
}
\arguments{
\item{object}{PMM3fit object}

\item{...}{Additional arguments (not used)}
}
\value{
Vector of coefficients
}
\description{
Extract coefficients from PMM3fit object
}
