% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_ts_methods.R
\name{coef,SARPMM2-method}
\alias{coef,SARPMM2-method}
\title{Extract coefficients from SARPMM2 object}
\usage{
\S4method{coef}{SARPMM2}(object, ...)
}
\arguments{
\item{object}{SARPMM2 object}

\item{...}{Additional arguments (not used)}
}
\value{
Named numeric vector of coefficients
}
\description{
Extract coefficients from SARPMM2 object
}
