% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_ts_methods.R
\name{coef,SMAPMM2-method}
\alias{coef,SMAPMM2-method}
\title{Extract coefficients from SMAPMM2 object}
\usage{
\S4method{coef}{SMAPMM2}(object, ...)
}
\arguments{
\item{object}{SMAPMM2 object}

\item{...}{Additional arguments (not used)}
}
\value{
Named vector of seasonal MA coefficients
}
\description{
Extract coefficients from SMAPMM2 object
}
