% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_ts_methods.R
\name{coef,TS2fit-method}
\alias{coef,TS2fit-method}
\title{Витягнути коефіцієнти з об'єкта TS2fit}
\usage{
\S4method{coef}{TS2fit}(object, ...)
}
\arguments{
\item{object}{Об'єкт TS2fit}

\item{...}{Додаткові аргументи (не використовуються)}
}
\value{
Іменований вектор коефіцієнтів
}
\description{
Витягнути коефіцієнти з об'єкта TS2fit
}
