% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm3_ts_methods.R
\name{coef,TS3fit-method}
\alias{coef,TS3fit-method}
\title{Extract coefficients from TS3fit object}
\usage{
\S4method{coef}{TS3fit}(object, ...)
}
\arguments{
\item{object}{TS3fit object}

\item{...}{Additional arguments (not used)}
}
\value{
Named vector of coefficients
}
\description{
Extract coefficients from TS3fit object
}
