% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_main.R
\name{compare_with_ols}
\alias{compare_with_ols}
\title{Porivniaty PMM2 z OLS}
\usage{
compare_with_ols(formula, data, pmm2_args = list())
}
\arguments{
\item{formula}{Formula modeli}

\item{data}{Freim danykh}

\item{pmm2_args}{Spysok arhumentiv dlia peredachi v lm_pmm2()}
}
\value{
Spysok z ob'iektamy pidhonky OLS ta PMM2
}
\description{
Porivniaty PMM2 z OLS
}
