% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_utils.R
\name{compute_moments}
\alias{compute_moments}
\title{Calculate moments and cumulants of error distribution}
\usage{
compute_moments(errors)
}
\arguments{
\item{errors}{numeric vector of errors}
}
\value{
list with moments, cumulants and theoretical variance reduction coefficient
}
\description{
Calculate moments and cumulants of error distribution
}
