% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_ts_design.R
\name{compute_ts_residuals}
\alias{compute_ts_residuals}
\title{Compute final residuals for time series models}
\usage{
compute_ts_residuals(coefs, model_info)
}
\arguments{
\item{coefs}{Estimated coefficients}

\item{model_info}{Model information}
}
\value{
Vector of residuals
}
\description{
Compute final residuals for time series models
}
\keyword{internal}
