% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm_utils.R
\name{create_ar_matrix}
\alias{create_ar_matrix}
\title{Create design matrix for AR models}
\usage{
create_ar_matrix(x, p)
}
\arguments{
\item{x}{centered time series}

\item{p}{AR order}
}
\value{
Design matrix with lagged values
}
\description{
Create design matrix for AR models
}
\keyword{internal}
