% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_ts_main.R
\name{create_residual_function}
\alias{create_residual_function}
\title{Create residual function for PMM2 optimization}
\usage{
create_residual_function(x, order, model_type, seasonal, include.mean)
}
\arguments{
\item{x}{Time series data}

\item{order}{Model order}

\item{model_type}{Model type}

\item{seasonal}{Seasonal specification}

\item{include.mean}{Include intercept}
}
\value{
Function that computes residuals given parameters
}
\description{
Create residual function for PMM2 optimization
}
\keyword{internal}
