% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_ts_design.R
\name{create_ts_design_matrix}
\alias{create_ts_design_matrix}
\title{Створення дизайн-матриці для часових рядів}
\usage{
create_ts_design_matrix(x, model_info, innovations = NULL)
}
\arguments{
\item{x}{Дані часового ряду}

\item{model_info}{Список з параметрами моделі}

\item{innovations}{Опціональні інновації/залишки для MA компонентів}
}
\value{
Список з дизайн-матрицею, змінною відгуку та іншими компонентами
}
\description{
Створення дизайн-матриці для часових рядів
}
\keyword{internal}
