% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm3_dispatch.R
\name{.g2_factor}
\alias{.g2_factor}
\title{PMM2 theoretical efficiency factor}
\usage{
.g2_factor(gamma3, gamma4)
}
\arguments{
\item{gamma3}{Skewness coefficient}

\item{gamma4}{Excess kurtosis coefficient}
}
\value{
Numeric scalar: g2 factor
}
\description{
PMM2 theoretical efficiency factor
}
\keyword{internal}
