% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm3_dispatch.R
\name{.g3_factor}
\alias{.g3_factor}
\title{PMM3 theoretical efficiency factor}
\usage{
.g3_factor(gamma4, gamma6)
}
\arguments{
\item{gamma4}{Excess kurtosis coefficient}

\item{gamma6}{Sixth-order cumulant coefficient}
}
\value{
Numeric scalar: g3 factor
}
\description{
PMM3 theoretical efficiency factor
}
\keyword{internal}
