% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm3_dispatch.R
\name{.residual_cumulants}
\alias{.residual_cumulants}
\title{Compute residual cumulants up to 6th order}
\usage{
.residual_cumulants(residuals)
}
\arguments{
\item{residuals}{Numeric vector of residuals}
}
\value{
Named list: m2, m3, m4, m6, gamma3, gamma4, gamma6
}
\description{
Compute residual cumulants up to 6th order
}
\keyword{internal}
