% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_ma_estimator.R
\name{estpmm_style_ma}
\alias{estpmm_style_ma}
\title{EstemPMM-style PMM2 estimator for MA models}
\usage{
estpmm_style_ma(x, q = 1, include.mean = TRUE, max_iter = 50, verbose = FALSE)
}
\arguments{
\item{x}{Time series}

\item{q}{MA order}

\item{include.mean}{Include intercept}

\item{max_iter}{Maximum PMM2 iterations}

\item{verbose}{Print diagnostics}
}
\value{
List with ma_coef, mean, innovations, convergence, method
}
\description{
EstemPMM-style PMM2 estimator for MA models
}
\keyword{internal}
