% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_ma_estimator.R
\name{estpmm_style_sma}
\alias{estpmm_style_sma}
\title{EstemPMM-style PMM2 estimator for SMA models}
\usage{
estpmm_style_sma(
  x,
  Q = 1,
  s = 4,
  include.mean = TRUE,
  max_iter = 50,
  verbose = FALSE
)
}
\arguments{
\item{x}{Time series}

\item{Q}{SMA order}

\item{s}{Seasonal period}

\item{include.mean}{Include intercept}

\item{max_iter}{Maximum PMM2 iterations}

\item{verbose}{Print diagnostics}
}
\value{
List with sma_coef, mean, innovations, convergence, method
}
\description{
EstemPMM-style PMM2 estimator for SMA models
}
\keyword{internal}
