% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_main.R
\name{fitted,PMM2fit-method}
\alias{fitted,PMM2fit-method}
\title{Vytiahnuty pidihnani znachennia z ob'iekta PMM2fit}
\usage{
\S4method{fitted}{PMM2fit}(object, data = NULL, ...)
}
\arguments{
\item{object}{Ob'iekt PMM2fit}

\item{data}{Neobov'iazkove dzherelo dlia rekonstruktsii modeli, yakshcho ob'iekt ne mistyt zberezhenykh danykh}

\item{...}{Dodatkovi arhumenty (ne vykorystovuiutsia)}
}
\value{
Vektor pidihnanykh znachen
}
\description{
Vytiahnuty pidihnani znachennia z ob'iekta PMM2fit
}
