% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm3_main.R
\name{fitted,PMM3fit-method}
\alias{fitted,PMM3fit-method}
\title{Extract fitted values from PMM3fit object}
\usage{
\S4method{fitted}{PMM3fit}(object, ...)
}
\arguments{
\item{object}{PMM3fit object}

\item{...}{Additional arguments (not used)}
}
\value{
Vector of fitted values
}
\description{
Extract fitted values from PMM3fit object
}
