% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_ts_methods.R
\name{fitted,TS2fit-method}
\alias{fitted,TS2fit-method}
\title{Vytiahnuty pidihnani znachennia z ob'iekta TS2fit}
\usage{
\S4method{fitted}{TS2fit}(object, ...)
}
\arguments{
\item{object}{Ob'iekt TS2fit}

\item{...}{Dodatkovi arhumenty (ne vykorystovuiutsia)}
}
\value{
Vektor pidihnanykh znachen
}
\description{
Vytiahnuty pidihnani znachennia z ob'iekta TS2fit
}
