% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm3_ts_methods.R
\name{fitted,TS3fit-method}
\alias{fitted,TS3fit-method}
\title{Extract fitted values from TS3fit object}
\usage{
\S4method{fitted}{TS3fit}(object, ...)
}
\arguments{
\item{object}{TS3fit object}

\item{...}{Additional arguments (not used)}
}
\value{
Vector of fitted values
}
\description{
Extract fitted values from TS3fit object
}
