% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_main.R
\name{fitted_values}
\alias{fitted_values}
\title{Dopomizhna funktsiia dlia vytiahnennia pidihnanykh znachen}
\usage{
fitted_values(object, data = NULL)
}
\arguments{
\item{object}{Ob'iekt PMM2fit}
}
\value{
Vektor pidihnanykh znachen
}
\description{
Dopomizhna funktsiia dlia vytiahnennia pidihnanykh znachen
}
\keyword{internal}
