% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm3_dispatch.R
\name{format.PMMdispatch}
\alias{format.PMMdispatch}
\title{Format method for PMMdispatch objects}
\usage{
\method{format}{PMMdispatch}(x, ...)
}
\arguments{
\item{x}{object of class \code{PMMdispatch}}

\item{...}{additional arguments (ignored)}
}
\value{
A character string summarising the selected method.
}
\description{
Format method for PMMdispatch objects
}
