% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm_ts.R
\name{get_ar_fitted}
\alias{get_ar_fitted}
\title{Calculate fitted values for AR models}
\usage{
get_ar_fitted(object)
}
\arguments{
\item{object}{A TS2fit object with model_type="ar"}
}
\value{
Vector of fitted values
}
\description{
Calculate fitted values for AR models
}
\keyword{internal}
