% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_ts_main.R
\name{get_classical_estimates}
\alias{get_classical_estimates}
\title{Get classical (MLE/CSS) estimates for initialization}
\usage{
get_classical_estimates(x, order, model_type, seasonal, include.mean)
}
\arguments{
\item{x}{Time series data}

\item{order}{Model order}

\item{model_type}{Model type}

\item{seasonal}{Seasonal specification}

\item{include.mean}{Include intercept}
}
\value{
Vector of initial parameter estimates
}
\description{
Get classical (MLE/CSS) estimates for initialization
}
\keyword{internal}
