% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_ts_design.R
\name{get_initial_estimates}
\alias{get_initial_estimates}
\title{Otrymaty pochatkovi otsinky parametriv dlia modelei chasovykh riadiv}
\usage{
get_initial_estimates(
  model_params,
  initial = NULL,
  method = "pmm2",
  verbose = FALSE
)
}
\arguments{
\item{model_params}{Perevireni parametry modeli z validate_ts_parameters}

\item{initial}{Optsionalno nadani korystuvachem pochatkovi otsinky}

\item{method}{Metod otsiniuvannia}

\item{verbose}{Vyvodyty detalnu informatsiiu}
}
\value{
Spysok, shcho mistyt:
\item{b_init}{vektor pochatkovykh koefitsiientiv AR/MA}
\item{x_mean}{otsinene serednie (iakshcho include.mean=TRUE)}
\item{innovations}{pochatkovi zalyshky/innovatsii}
\item{x_centered}{tsentrovanyi (abo dyferentsiiovanyi + tsentrovanyi) riad}
\item{m2}{druhyi tsentralnyi moment pochatkovykh zalyshkiv}
\item{m3}{tretii tsentralnyi moment pochatkovykh zalyshkiv}
\item{m4}{chetvertyi tsentralnyi moment pochatkovykh zalyshkiv}
}
\description{
Otrymaty pochatkovi otsinky parametriv dlia modelei chasovykh riadiv
}
\keyword{internal}
