% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarimax_wrapper.R
\name{get_sarimax_jacobian}
\alias{get_sarimax_jacobian}
\title{Calculate SARIMAX Jacobian (Numerical)}
\usage{
get_sarimax_jacobian(theta, ...)
}
\arguments{
\item{theta}{Parameters}

\item{...}{Arguments passed to get_sarimax_residuals}
}
\value{
Jacobian matrix (n x p)
}
\description{
Calculate SARIMAX Jacobian (Numerical)
}
