% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarimax_wrapper.R
\name{get_sarimax_residuals}
\alias{get_sarimax_residuals}
\title{Calculate SARIMAX Residuals}
\usage{
get_sarimax_residuals(
  theta,
  y,
  xreg = NULL,
  order = c(0, 0, 0),
  seasonal = list(order = c(0, 0, 0), period = NA),
  include.mean = TRUE
)
}
\arguments{
\item{theta}{Combined vector of parameters (AR, MA, SAR, SMA, Intercept, Regressors)}

\item{y}{Time series data}

\item{xreg}{Exogenous regressors (optional)}

\item{order}{ARIMA order c(p, d, q)}

\item{seasonal}{Seasonal order c(P, D, Q)}

\item{include.mean}{Boolean, whether to include mean/intercept}

\item{period}{Seasonal period}
}
\value{
Vector of residuals
}
\description{
Calculate SARIMAX Residuals
}
