% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_ts_design.R
\name{get_yw_estimates}
\alias{get_yw_estimates}
\title{Get Yule-Walker estimates for AR(p)}
\usage{
get_yw_estimates(x, p)
}
\arguments{
\item{x}{Numeric vector}

\item{p}{Integer value of AR order}
}
\value{
Numeric vector of length p (AR coefficients)
}
\description{
Get Yule-Walker estimates for AR(p)
}
\keyword{internal}
