% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_ts_main.R
\name{linearized_pmm2_wrapper}
\alias{linearized_pmm2_wrapper}
\title{Wrapper for linearized PMM2 approach (MA/SMA models)}
\usage{
linearized_pmm2_wrapper(
  x,
  order,
  model_type,
  seasonal,
  include.mean,
  max_iter,
  tol,
  verbose
)
}
\arguments{
\item{x}{Time series data}

\item{order}{Model order}

\item{model_type}{Model type}

\item{seasonal}{Seasonal specification}

\item{include.mean}{Include intercept}

\item{max_iter}{Maximum iterations}

\item{tol}{Convergence tolerance}

\item{verbose}{Print progress}
}
\value{
PMM2 estimation results
}
\description{
Wrapper for linearized PMM2 approach (MA/SMA models)
}
\keyword{internal}
