% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_main.R
\name{logLik,PMM2fit-method}
\alias{logLik,PMM2fit-method}
\title{Extract log-likelihood from PMM2fit object}
\usage{
\S4method{logLik}{PMM2fit}(object, ...)
}
\arguments{
\item{object}{PMM2fit object}

\item{...}{Additional arguments (not used)}
}
\value{
Object of class \code{logLik}
}
\description{
Returns a Gaussian approximate log-likelihood, consistent with the AIC method.
}
