% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm3_main.R
\name{logLik,PMM3fit-method}
\alias{logLik,PMM3fit-method}
\title{Extract log-likelihood from PMM3fit object}
\usage{
\S4method{logLik}{PMM3fit}(object, ...)
}
\arguments{
\item{object}{PMM3fit object}

\item{...}{Additional arguments (not used)}
}
\value{
Object of class \code{logLik}
}
\description{
Returns a Gaussian approximate log-likelihood, consistent with the AIC method.
}
