% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_ts_methods.R
\name{logLik,TS2fit-method}
\alias{logLik,TS2fit-method}
\title{Extract log-likelihood from TS2fit object}
\usage{
\S4method{logLik}{TS2fit}(object, ...)
}
\arguments{
\item{object}{TS2fit (or subclass) object}

\item{...}{Additional arguments (not used)}
}
\value{
Object of class \code{logLik}
}
\description{
Returns a Gaussian approximate log-likelihood, consistent with the AIC method.
}
