% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm3_ts_methods.R
\name{logLik,TS3fit-method}
\alias{logLik,TS3fit-method}
\title{Extract log-likelihood from TS3fit object}
\usage{
\S4method{logLik}{TS3fit}(object, ...)
}
\arguments{
\item{object}{TS3fit (or subclass) object}

\item{...}{Additional arguments (not used)}
}
\value{
Object of class \code{logLik}
}
\description{
Returns a Gaussian approximate log-likelihood, consistent with the AIC method.
}
