% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_main.R
\name{logLik.PMM2fit}
\alias{logLik.PMM2fit}
\title{Extract log-likelihood from PMM2fit object}
\usage{
\method{logLik}{PMM2fit}(object, ...)
}
\arguments{
\item{object}{PMM2fit object}

\item{...}{Additional arguments (not used)}
}
\value{
Object of class \code{logLik}
}
\description{
Returns a Gaussian approximate log-likelihood. Defined as an S3 method
so that \code{stats::AIC} and \code{stats::BIC} (which dispatch
\code{logLik} via \code{UseMethod}) work without explicit S4 methods.
}
