% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm3_main.R
\name{logLik.PMM3fit}
\alias{logLik.PMM3fit}
\title{Extract log-likelihood from PMM3fit object}
\usage{
\method{logLik}{PMM3fit}(object, ...)
}
\arguments{
\item{object}{PMM3fit object}

\item{...}{Additional arguments (not used)}
}
\value{
Object of class \code{logLik}
}
\description{
Returns a Gaussian approximate log-likelihood. Defined as an S3 method
so that \code{stats::AIC} and \code{stats::BIC} (which dispatch
\code{logLik} via \code{UseMethod}) work without explicit S4 methods.
}
