% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm3_ts_methods.R
\name{logLik.TS3fit}
\alias{logLik.TS3fit}
\title{Extract log-likelihood from TS3fit object}
\usage{
\method{logLik}{TS3fit}(object, ...)
}
\arguments{
\item{object}{TS3fit (or subclass) object}

\item{...}{Additional arguments (not used)}
}
\value{
Object of class \code{logLik}
}
\description{
Returns a Gaussian approximate log-likelihood. Defined as an S3 method
so that \code{stats::AIC} and \code{stats::BIC} dispatch correctly via
\code{UseMethod("logLik")}.
}
