% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_ma_estimator.R
\name{ma_build_design}
\alias{ma_build_design}
\title{Build design matrix for MA}
\usage{
ma_build_design(intercept, residuals, x, q)
}
\arguments{
\item{intercept}{Intercept (mean)}

\item{residuals}{Initial residuals from CSS}

\item{x}{Time series}

\item{q}{MA order}
}
\value{
List with X (design matrix) and y (response)
}
\description{
Build design matrix for MA
}
\keyword{internal}
