% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_ma_estimator.R
\name{ma_compute_innovations}
\alias{ma_compute_innovations}
\title{Compute MA innovations (forward recursion)}
\usage{
ma_compute_innovations(x, theta, q)
}
\arguments{
\item{x}{Time series (centered)}

\item{theta}{MA coefficients}

\item{q}{MA order}
}
\value{
Vector of innovations
}
\description{
Compute MA innovations (forward recursion)
}
\keyword{internal}
