% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm_ts.R
\name{ma_predictions}
\alias{ma_predictions}
\title{Generate MA model predictions}
\usage{
ma_predictions(innovations, ma_coef, n.ahead)
}
\arguments{
\item{innovations}{Vector of innovations}

\item{ma_coef}{Vector of MA coefficients}

\item{n.ahead}{Number of steps to predict}
}
\value{
Vector of predictions
}
\description{
Generate MA model predictions
}
\keyword{internal}
