% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_ma_estimator.R
\name{ma_sma_build_design}
\alias{ma_sma_build_design}
\title{Build design matrix for combined MA+SMA model}
\usage{
ma_sma_build_design(intercept, residuals, x, q, Q, s, multiplicative = TRUE)
}
\arguments{
\item{intercept}{Intercept (mean)}

\item{residuals}{Initial residuals from CSS}

\item{x}{Time series}

\item{q}{MA order}

\item{Q}{SMA order}

\item{s}{Seasonal period}

\item{multiplicative}{Logical, if TRUE use multiplicative SARIMA model}
}
\value{
List with X (design matrix) and y (response)
}
\description{
Build design matrix for combined MA+SMA model
}
\keyword{internal}
