% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_ma_estimator.R
\name{ma_sma_compute_innovations}
\alias{ma_sma_compute_innovations}
\title{Compute combined MA+SMA innovations (forward recursion)}
\usage{
ma_sma_compute_innovations(
  x,
  theta_ma,
  theta_sma,
  q,
  Q,
  s,
  multiplicative = TRUE
)
}
\arguments{
\item{x}{Time series (centered)}

\item{theta_ma}{MA coefficients (length q)}

\item{theta_sma}{SMA coefficients (length Q)}

\item{q}{MA order}

\item{Q}{SMA order}

\item{s}{Seasonal period}

\item{multiplicative}{Logical, if TRUE use multiplicative SARIMA model}
}
\value{
Vector of innovations
}
\description{
Compute combined MA+SMA innovations (forward recursion)
}
\keyword{internal}
