% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_ma_estimator.R
\name{ma_solve_pmm2}
\alias{ma_solve_pmm2}
\title{PMM2 solver (EstemPMM formula)}
\usage{
ma_solve_pmm2(
  b_init,
  X,
  Y,
  m2,
  m3,
  m4,
  max_iter = 50,
  tol = 1e-06,
  verbose = FALSE
)
}
\arguments{
\item{b_init}{Initial coefficients vector (intercept and MA coefficients)}

\item{X}{Design matrix}

\item{Y}{Response vector}

\item{m2}{Second central moment}

\item{m3}{Third central moment}

\item{m4}{Fourth central moment}

\item{max_iter}{Maximum iterations}

\item{tol}{Convergence tolerance}

\item{verbose}{Print diagnostics}
}
\value{
List with coefficients, convergence, iterations
}
\description{
PMM2 solver (EstemPMM formula)
}
\keyword{internal}
