% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_main.R
\name{nobs,PMM2fit-method}
\alias{nobs,PMM2fit-method}
\title{Number of observations in PMM2fit object}
\usage{
\S4method{nobs}{PMM2fit}(object, ...)
}
\arguments{
\item{object}{PMM2fit object}

\item{...}{Additional arguments (not used)}
}
\value{
Integer number of observations
}
\description{
Number of observations in PMM2fit object
}
