% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm3_main.R
\name{nobs,PMM3fit-method}
\alias{nobs,PMM3fit-method}
\title{Number of observations in PMM3fit object}
\usage{
\S4method{nobs}{PMM3fit}(object, ...)
}
\arguments{
\item{object}{PMM3fit object}

\item{...}{Additional arguments (not used)}
}
\value{
Integer number of observations
}
\description{
Number of observations in PMM3fit object
}
