% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_ts_methods.R
\name{nobs,TS2fit-method}
\alias{nobs,TS2fit-method}
\title{Number of observations in TS2fit object}
\usage{
\S4method{nobs}{TS2fit}(object, ...)
}
\arguments{
\item{object}{TS2fit (or subclass) object}

\item{...}{Additional arguments (not used)}
}
\value{
Integer effective sample size
}
\description{
Returns the effective sample size (length of the residual vector).
}
