% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm3_ts_methods.R
\name{nobs,TS3fit-method}
\alias{nobs,TS3fit-method}
\title{Number of observations in TS3fit object}
\usage{
\S4method{nobs}{TS3fit}(object, ...)
}
\arguments{
\item{object}{TS3fit (or subclass) object}

\item{...}{Additional arguments (not used)}
}
\value{
Integer effective sample size
}
\description{
Number of observations in TS3fit object
}
