% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_main.R
\name{plot,PMM2fit,missing-method}
\alias{plot,PMM2fit,missing-method}
\title{Pobuduvaty diahnostychni hrafiky dlia ob'iekta PMM2fit}
\usage{
\S4method{plot}{PMM2fit,missing}(x, y, which = 1:4, ...)
}
\arguments{
\item{x}{Ob'iekt PMM2fit}

\item{y}{Ne vykorystovuietsia (sumisnist z generic)}

\item{which}{Nabir hrafikiv dlia vidobrazhennia (znachennia 1-4)}

\item{...}{Dodatkovi arhumenty, shcho peredaiutsia hrafichnym funktsiiam}
}
\value{
Nevydymo povertaie vkhidnyi ob'iekt
}
\description{
Pobuduvaty diahnostychni hrafiky dlia ob'iekta PMM2fit
}
