% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm3_main.R
\name{plot,PMM3fit,missing-method}
\alias{plot,PMM3fit,missing-method}
\title{Plot diagnostic plots for PMM3fit object}
\usage{
\S4method{plot}{PMM3fit,missing}(x, y, which = 1:4, ...)
}
\arguments{
\item{x}{PMM3fit object}

\item{y}{Not used (compatibility with generic)}

\item{which}{Set of plots to display (values 1-4)}

\item{...}{Additional arguments passed to plotting functions}
}
\value{
Invisibly returns the input object
}
\description{
Plot diagnostic plots for PMM3fit object
}
