% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_ts_methods.R
\name{plot,TS2fit,missing-method}
\alias{plot,TS2fit,missing-method}
\title{Pobuduvaty diahnostychni hrafiky dlia ob'iektiv TS2fit}
\usage{
\S4method{plot}{TS2fit,missing}(x, y, which = c(1:4), ...)
}
\arguments{
\item{x}{Ob'iekt TS2fit}

\item{y}{Ne vykorystovuietsia (dlia sumisnosti metodu S4)}

\item{which}{Tsilochyselnyi vektor, shcho vkazuie, iaki hrafiky vyrobliaty}

\item{...}{dodatkovi arhumenty, peredani funktsiiam pobudovy hrafikiv}
}
\value{
Nevydymo povertaie x
}
\description{
Pobuduvaty diahnostychni hrafiky dlia ob'iektiv TS2fit
}
