% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm3_ts_methods.R
\name{plot,TS3fit,missing-method}
\alias{plot,TS3fit,missing-method}
\title{Plot diagnostic plots for TS3fit object}
\usage{
\S4method{plot}{TS3fit,missing}(x, y, which = 1:4, ...)
}
\arguments{
\item{x}{TS3fit object}

\item{y}{Not used}

\item{which}{Set of plots to display (values 1-4)}

\item{...}{Additional arguments}
}
\value{
Invisibly returns the input object
}
\description{
Plot diagnostic plots for TS3fit object
}
