% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_inference.R
\name{plot_pmm2_bootstrap}
\alias{plot_pmm2_bootstrap}
\title{Plot bootstrap distributions for PMM2 fit}
\usage{
plot_pmm2_bootstrap(object, coefficients = NULL)
}
\arguments{
\item{object}{Result from pmm2_inference}

\item{coefficients}{Which coefficients to plot, defaults to all}
}
\value{
Invisibly returns histogram information
}
\description{
Plot bootstrap distributions for PMM2 fit
}
