% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm_package.R
\docType{package}
\name{pmm2}
\alias{EstemPMM}
\alias{EstemPMM-package}
\alias{pmm2}
\title{PMM2: Поліноміальний метод максимізації для робастної регресії та часових рядів}
\description{
Пакет PMM2 пропонує робастні методи для оцінювання параметрів лінійних
моделей та моделей часових рядів, робастних до негаусівських помилок.
}
\section{Функції лінійної регресії}{


\code{\link{lm_pmm2}} - Підгонка лінійної моделі за допомогою PMM2

\code{\link{compare_with_ols}} - Порівняння PMM2 з OLS
}

\section{Функції часових рядів}{


\code{\link{ts_pmm2}} - Загальна функція для підгонки моделей часових рядів за допомогою PMM2

\code{\link{ar_pmm2}} - Підгонка AR моделей

\code{\link{ma_pmm2}} - Підгонка MA моделей

\code{\link{arma_pmm2}} - Підгонка ARMA моделей

\code{\link{arima_pmm2}} - Підгонка ARIMA моделей

\code{\link{compare_ts_methods}} - Порівняння PMM2 з класичними методами
}

\section{Статистичний висновок}{


\code{\link{pmm2_inference}} - Бутстреп-висновок для лінійних моделей

\code{\link{ts_pmm2_inference}} - Бутстреп-висновок для моделей часових рядів
}

\section{Утиліти}{


\code{\link{pmm_skewness}} - Обчислення асиметрії

\code{\link{pmm_kurtosis}} - Обчислення ексцесу

\code{\link{compute_moments}} - Обчислення моментів та кумулянтів
}

\author{
\strong{Maintainer}: Serhii Zabolotnii \email{zabolotniua@gmail.com}

}
