% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_utils.R
\name{pmm2_variance_factor}
\alias{pmm2_variance_factor}
\title{Обчислити теоретичні коефіцієнти асиметрії, ексцесу та фактор зменшення дисперсії}
\usage{
pmm2_variance_factor(m2, m3, m4)
}
\arguments{
\item{m2, m3, m4}{центральні моменти другого, третього та четвертого порядків}
}
\value{
Список з полями \code{c3}, \code{c4} та \code{g}
}
\description{
Обчислити теоретичні коефіцієнти асиметрії, ексцесу та фактор зменшення дисперсії
}
