% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_classes.R, R/pmm2_package.R
\docType{class}
\name{PMM2fit-class}
\alias{PMM2fit-class}
\title{Клас S4 для зберігання результатів PMM2 регресійної моделі}
\description{
Клас для зберігання результатів оцінювання лінійних моделей за допомогою PMM2
}
\section{Slots}{

\describe{
\item{\code{coefficients}}{числовий вектор оцінених параметрів}

\item{\code{residuals}}{числовий вектор кінцевих залишків}

\item{\code{m2}}{числовий другий центральний момент початкових залишків}

\item{\code{m3}}{числовий третій центральний момент початкових залишків}

\item{\code{m4}}{числовий четвертий центральний момент початкових залишків}

\item{\code{convergence}}{логічний або цілий код, що вказує, чи алгоритм збігся}

\item{\code{iterations}}{числова кількість виконаних ітерацій}

\item{\code{call}}{оригінальний виклик функції}
}}

\section{Слоти}{

\describe{
\item{coefficients}{Оцінені коефіцієнти}
\item{residuals}{Кінцеві залишки}
\item{m2}{Другий центральний момент}
\item{m3}{Третій центральний момент}
\item{m4}{Четвертий центральний момент}
\item{convergence}{Статус збіжності}
\item{iterations}{Кількість виконаних ітерацій}
\item{call}{Оригінальний виклик}
}
}

