% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm3_classes.R
\docType{class}
\name{PMM3fit-class}
\alias{PMM3fit-class}
\title{S4 class for PMM3 regression fit results}
\description{
Concrete class returned by \code{\link{lm_pmm3}}. Inherits all slots
from \code{\linkS4class{BasePMM3}}, which in turn inherits common
slots from \code{\linkS4class{PMMfit}}.
}
