% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm_unified_api.R
\name{pmm_ar}
\alias{pmm_ar}
\title{Fit an AR model with the polynomial maximization method}
\usage{
pmm_ar(x, order, method = c("pmm2", "pmm3"), ...)
}
\arguments{
\item{x}{univariate time series or numeric vector.}

\item{order}{non-negative integer giving the autoregressive order.}

\item{method}{One of \code{"pmm2"} or \code{"pmm3"}.}

\item{...}{Additional arguments forwarded to \code{\link{ar_pmm2}}
or \code{\link{ar_pmm3}}.}
}
\value{
An S4 object of class \code{\linkS4class{ARPMM2}} (for
\code{method = "pmm2"}) or \code{\linkS4class{ARPMM3}}.
}
\description{
Unified entry point for PMM2 and PMM3 autoregressive estimation.
}
