% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm_unified_api.R
\name{pmm_arima}
\alias{pmm_arima}
\title{Fit an ARIMA model with the polynomial maximization method}
\usage{
pmm_arima(x, order, method = c("pmm2", "pmm3"), ...)
}
\arguments{
\item{x}{univariate time series or numeric vector.}

\item{order}{integer vector \code{c(p, d, q)} giving the AR order,
degree of differencing, and MA order.}

\item{method}{One of \code{"pmm2"} or \code{"pmm3"}.}

\item{...}{Additional arguments forwarded to \code{\link{arima_pmm2}}
or \code{\link{arima_pmm3}}.}
}
\value{
An S4 object of class \code{\linkS4class{ARIMAPMM2}} or
\code{\linkS4class{ARIMAPMM3}}.
}
\description{
Unified entry point for PMM2 and PMM3 ARIMA estimation. This is the
recommended primary API for non-Gaussian ARIMA modelling and the
interface featured in the JSS submission accompanying the package.
}
\examples{
set.seed(2026)
x <- arima.sim(list(ar = 0.6), n = 200,
               rand.gen = function(k) rgamma(k, 2, 1) - 2)
pmm_arima(x, order = c(1, 0, 0), method = "pmm2")

}
