% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm_unified_api.R
\name{pmm_arma}
\alias{pmm_arma}
\title{Fit an ARMA model with the polynomial maximization method}
\usage{
pmm_arma(x, order, method = c("pmm2", "pmm3"), ...)
}
\arguments{
\item{x}{univariate time series or numeric vector.}

\item{order}{integer vector \code{c(p, q)} giving the AR and MA orders.}

\item{method}{One of \code{"pmm2"} or \code{"pmm3"}.}

\item{...}{Additional arguments forwarded to \code{\link{arma_pmm2}}
or \code{\link{arma_pmm3}}.}
}
\value{
An S4 object of class \code{\linkS4class{ARMAPMM2}} or
\code{\linkS4class{ARMAPMM3}}.
}
\description{
Unified entry point for PMM2 and PMM3 ARMA estimation.
}
