% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_utils.R
\name{pmm_kurtosis}
\alias{pmm_kurtosis}
\title{Calculate kurtosis from data}
\usage{
pmm_kurtosis(x, excess = TRUE)
}
\arguments{
\item{x}{numeric vector}

\item{excess}{logical, whether to return excess kurtosis (kurtosis - 3)}
}
\value{
Kurtosis value
}
\description{
Calculate kurtosis from data
}
