% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm_unified_api.R
\name{pmm_ma}
\alias{pmm_ma}
\title{Fit an MA model with the polynomial maximization method}
\usage{
pmm_ma(x, order, method = c("pmm2", "pmm3"), ...)
}
\arguments{
\item{x}{univariate time series or numeric vector.}

\item{order}{non-negative integer giving the moving-average order.}

\item{method}{One of \code{"pmm2"} or \code{"pmm3"}.}

\item{...}{Additional arguments forwarded to \code{\link{ma_pmm2}}
or \code{\link{ma_pmm3}}.}
}
\value{
An S4 object of class \code{\linkS4class{MAPMM2}} or
\code{\linkS4class{MAPMM3}}.
}
\description{
Unified entry point for PMM2 and PMM3 moving-average estimation.
}
