% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm_unified_api.R
\name{pmm_sarima}
\alias{pmm_sarima}
\title{Fit a seasonal ARIMA model with the polynomial maximization method}
\usage{
pmm_sarima(x, ..., method = c("pmm2", "pmm3"))
}
\arguments{
\item{x}{univariate time series or numeric vector.}

\item{...}{Arguments forwarded to \code{\link{sarima_pmm2}} (notably
\code{order} and \code{seasonal}).}

\item{method}{Currently only \code{"pmm2"} is supported.}
}
\value{
An S4 object of class \code{\linkS4class{SARIMAPMM2}}.
}
\description{
Unified entry point for SARIMA-PMM2 estimation. PMM3 seasonal
estimation is not yet implemented; \code{method = "pmm3"} therefore
raises an error. All arguments other than \code{method} are
forwarded verbatim to \code{\link{sarima_pmm2}} -- see that function
for the seasonal order specification.
}
