% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_utils.R
\name{pmm_skewness}
\alias{pmm_skewness}
\title{Calculate skewness from data}
\usage{
pmm_skewness(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
Skewness value
}
\description{
Calculate skewness from data
}
