% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm3_main.R
\name{predict,PMM3fit-method}
\alias{predict,PMM3fit-method}
\title{Predict method for PMM3fit objects}
\usage{
\S4method{predict}{PMM3fit}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{PMM3fit object}

\item{newdata}{Data frame with predictor variables}

\item{...}{Additional arguments (not used)}
}
\value{
Numeric vector of predicted values
}
\description{
Predict method for PMM3fit objects
}
