% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_ts_methods.R
\name{predict,TS2fit-method}
\alias{predict,TS2fit-method}
\title{Prediction method for TS2fit objects}
\usage{
\S4method{predict}{TS2fit}(object, n.ahead = 1, ...)
}
\arguments{
\item{object}{TS2fit object}

\item{n.ahead}{Number of steps ahead for prediction}

\item{...}{additional arguments (not used)}
}
\value{
Vector or list of predictions, depending on model type
}
\description{
Prediction method for TS2fit objects
}
