% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm3_ts_methods.R
\name{predict,TS3fit-method}
\alias{predict,TS3fit-method}
\title{Predict method for TS3fit objects}
\usage{
\S4method{predict}{TS3fit}(object, n.ahead = 1, ...)
}
\arguments{
\item{object}{TS3fit object}

\item{n.ahead}{Integer: number of steps ahead to forecast}

\item{...}{Additional arguments (not used)}
}
\value{
Numeric vector of predicted values
}
\description{
Predict method for TS3fit objects
}
