% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm3_dispatch.R
\name{print.PMMdispatch}
\alias{print.PMMdispatch}
\title{Print method for PMMdispatch objects}
\usage{
\method{print}{PMMdispatch}(x, ...)
}
\arguments{
\item{x}{object of class \code{PMMdispatch} returned by
\code{\link{pmm_dispatch}}}

\item{...}{additional arguments (ignored)}
}
\value{
The object \code{x}, invisibly.
}
\description{
Print method for PMMdispatch objects
}
