% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_main.R
\name{residuals,PMM2fit-method}
\alias{residuals,PMM2fit-method}
\title{Vytiahnuty zalyshky z ob'iekta PMM2fit}
\usage{
\S4method{residuals}{PMM2fit}(object, ...)
}
\arguments{
\item{object}{Ob'iekt PMM2fit}

\item{...}{Dodatkovi arhumenty (ne vykorystovuiutsia)}
}
\value{
Vektor zalyshkiv
}
\description{
Vytiahnuty zalyshky z ob'iekta PMM2fit
}
