% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm3_main.R
\name{residuals,PMM3fit-method}
\alias{residuals,PMM3fit-method}
\title{Extract residuals from PMM3fit object}
\usage{
\S4method{residuals}{PMM3fit}(object, ...)
}
\arguments{
\item{object}{PMM3fit object}

\item{...}{Additional arguments (not used)}
}
\value{
Vector of residuals
}
\description{
Extract residuals from PMM3fit object
}
