% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm3_ts_methods.R
\name{residuals,TS3fit-method}
\alias{residuals,TS3fit-method}
\title{Extract residuals from TS3fit object}
\usage{
\S4method{residuals}{TS3fit}(object, ...)
}
\arguments{
\item{object}{TS3fit object}

\item{...}{Additional arguments (not used)}
}
\value{
Vector of residuals
}
\description{
Extract residuals from TS3fit object
}
