% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_classes.R
\docType{class}
\name{SARPMM2-class}
\alias{SARPMM2-class}
\title{S4 class for Seasonal AR model results with PMM2}
\description{
This class stores the results of fitting a Seasonal Autoregressive (SAR)
model using the PMM2 method. It extends the TS2fit class with additional
slots specific to seasonal models.
}
\details{
The SARPMM2 class represents fitted SAR models of the form
\deqn{y_t = \phi_1 y_{t-1} + \dots + \phi_p y_{t-p} +
            \Phi_1 y_{t-s} + \dots + \Phi_P y_{t-Ps} + \epsilon_t.}

Where:
\itemize{
\item p is the non-seasonal AR order
\item P is the seasonal AR order
\item s is the seasonal period
}
}
\section{Slots}{

\describe{
\item{\code{coefficients}}{Numeric vector of estimated parameters (AR and SAR coefficients)}

\item{\code{residuals}}{Numeric vector of residuals/innovations}

\item{\code{m2}}{Second central moment of residuals}

\item{\code{m3}}{Third central moment of residuals}

\item{\code{m4}}{Fourth central moment of residuals}

\item{\code{convergence}}{Logical, whether PMM2 algorithm converged}

\item{\code{iterations}}{Integer, number of iterations performed}

\item{\code{call}}{Original function call}

\item{\code{model_type}}{Character, model type identifier ("sar")}

\item{\code{intercept}}{Numeric, intercept/mean term}

\item{\code{original_series}}{Numeric vector, original time series data}

\item{\code{order}}{List with model specification: list(ar, sar, period)
\itemize{
\item ar: Non-seasonal AR order (p)
\item sar: Seasonal AR order (P)
\item period: Seasonal period (s)
}}
}}

\seealso{
\code{\link{sar_pmm2}} for fitting SAR models
}
