% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm_show_methods.R
\name{show,TS2fit-method}
\alias{show,TS2fit-method}
\title{Show method for TS2fit objects (and subclasses)}
\usage{
\S4method{show}{TS2fit}(object)
}
\arguments{
\item{object}{object inheriting from \code{TS2fit}}
}
\value{
The object (invisibly).
}
\description{
Inherited by ARPMM2, MAPMM2, ARMAPMM2, ARIMAPMM2 and the seasonal
SARPMM2/SMAPMM2/SARMAPMM2/SARIMAPMM2 subclasses. Use
\code{\link{summary}} for full diagnostics.
}
