% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm_show_methods.R
\name{show,TS3fit-method}
\alias{show,TS3fit-method}
\title{Show method for TS3fit objects (and subclasses)}
\usage{
\S4method{show}{TS3fit}(object)
}
\arguments{
\item{object}{object inheriting from \code{TS3fit}}
}
\value{
The object (invisibly).
}
\description{
Inherited by ARPMM3, MAPMM3, ARMAPMM3, ARIMAPMM3. Use
\code{\link{summary}} for full diagnostics.
}
