% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_ma_estimator.R
\name{sma_build_design}
\alias{sma_build_design}
\title{Build design matrix for SMA}
\usage{
sma_build_design(intercept, residuals, x, Q, s)
}
\arguments{
\item{intercept}{Intercept (mean)}

\item{residuals}{Initial residuals from CSS}

\item{x}{Time series}

\item{Q}{SMA order}

\item{s}{Seasonal period}
}
\value{
List with X (design matrix) and y (response)
}
\description{
Build design matrix for SMA
}
\keyword{internal}
