% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_ts_main.R
\name{sma_css_fit}
\alias{sma_css_fit}
\title{CSS fit for Seasonal MA model}
\usage{
sma_css_fit(x, Q, s, include_mean = TRUE, verbose = FALSE)
}
\arguments{
\item{x}{Time series data}

\item{Q}{Seasonal MA order}

\item{s}{Seasonal period}

\item{include_mean}{Include intercept}

\item{verbose}{Print diagnostics}
}
\value{
List with coefficients, intercept, residuals, convergence info
}
\description{
CSS fit for Seasonal MA model
}
\keyword{internal}
