% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_ts_main.R
\name{sma_pmm2_fit}
\alias{sma_pmm2_fit}
\title{Fit Seasonal MA model using PMM2}
\usage{
sma_pmm2_fit(x, Q, s, css_fit, max_iter = 50, tol = 1e-06, verbose = FALSE)
}
\arguments{
\item{x}{Time series data}

\item{Q}{Seasonal MA order}

\item{s}{Seasonal period}

\item{css_fit}{Initial CSS fit}

\item{max_iter}{Maximum iterations}

\item{tol}{Convergence tolerance}

\item{verbose}{Print progress}
}
\value{
List with coefficients, intercept, innovations, convergence info
}
\description{
Fit Seasonal MA model using PMM2
}
\keyword{internal}
