% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_classes.R
\name{summary,PMM2fit-method}
\alias{summary,PMM2fit-method}
\title{Uzahalnenyi metod summary dlia ob'iektiv PMM2fit}
\usage{
\S4method{summary}{PMM2fit}(object, formula = NULL, data = NULL, B = 100, ...)
}
\arguments{
\item{object}{ob'iekt klasu "PMM2fit"}

\item{formula}{(optsionalno) formula, vykorystana dlia modeli}

\item{data}{(optsionalno) vykorystani dani}

\item{B}{kilkist butstrep-replikatsii dlia statystychnoho vysnovku}

\item{...}{dodatkovi arhumenty (ne vykorystovuiutsia)}
}
\value{
Vyvodyt reziume na konsol; povertaie ob'iekt (nevydymo).
}
\description{
Uzahalnenyi metod summary dlia ob'iektiv PMM2fit
}
