% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm3_main.R
\name{summary,PMM3fit-method}
\alias{summary,PMM3fit-method}
\title{Summary method for PMM3fit objects}
\usage{
\S4method{summary}{PMM3fit}(object, ...)
}
\arguments{
\item{object}{PMM3fit object}

\item{...}{Additional arguments (not used)}
}
\value{
Prints summary to console; returns object (invisibly)
}
\description{
Summary method for PMM3fit objects
}
