% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_classes.R
\name{summary,SARIMAPMM2-method}
\alias{summary,SARIMAPMM2-method}
\title{Generic summary method for SARIMAPMM2 objects}
\usage{
\S4method{summary}{SARIMAPMM2}(object, ...)
}
\arguments{
\item{object}{object of class "SARIMAPMM2"}

\item{...}{additional arguments (not used)}
}
\value{
Prints summary to console; returns object (invisibly).
}
\description{
Generic summary method for SARIMAPMM2 objects
}
