% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_classes.R
\name{summary,SARMAPMM2-method}
\alias{summary,SARMAPMM2-method}
\title{Generic summary method for SARMAPMM2 objects}
\usage{
\S4method{summary}{SARMAPMM2}(object, ...)
}
\arguments{
\item{object}{object of class "SARMAPMM2"}

\item{...}{additional arguments (not used)}
}
\value{
Prints summary to console; returns object (invisibly).
}
\description{
Generic summary method for SARMAPMM2 objects
}
