% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_ts_methods.R
\name{summary,SARPMM2-method}
\alias{summary,SARPMM2-method}
\title{Summary method for SARPMM2 objects}
\usage{
\S4method{summary}{SARPMM2}(object, ...)
}
\arguments{
\item{object}{SARPMM2 object}

\item{...}{Additional arguments (not used)}
}
\value{
Invisibly returns the object
}
\description{
Summary method for SARPMM2 objects
}
