% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_ts_methods.R
\name{summary,SMAPMM2-method}
\alias{summary,SMAPMM2-method}
\title{Summary method for SMAPMM2 objects}
\usage{
\S4method{summary}{SMAPMM2}(object, ...)
}
\arguments{
\item{object}{SMAPMM2 object}

\item{...}{Additional arguments (not used)}
}
\value{
Invisibly returns the object
}
\description{
Summary method for SMAPMM2 objects
}
