% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_classes.R
\name{summary,TS2fit-method}
\alias{summary,TS2fit-method}
\title{Generic summary method for TS2fit objects}
\usage{
\S4method{summary}{TS2fit}(object, ...)
}
\arguments{
\item{object}{object of class "TS2fit" or subclass}

\item{...}{additional arguments (not used)}
}
\value{
Prints summary to console; returns object (invisibly).
}
\description{
Generic summary method for TS2fit objects
}
