% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm3_ts_methods.R
\name{summary,TS3fit-method}
\alias{summary,TS3fit-method}
\title{Summary method for TS3fit objects}
\usage{
\S4method{summary}{TS3fit}(object, ...)
}
\arguments{
\item{object}{TS3fit object}

\item{...}{Additional arguments (not used)}
}
\value{
Prints summary to console; returns object (invisibly)
}
\description{
Summary method for TS3fit objects
}
