% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_ts_design.R
\name{update_ma_innovations}
\alias{update_ma_innovations}
\title{Update MA model innovations}
\usage{
update_ma_innovations(x, ma_coef)
}
\arguments{
\item{x}{Centered time series}

\item{ma_coef}{Vector of MA coefficients}
}
\value{
Vector of innovations
}
\description{
Update MA model innovations
}
\keyword{internal}
