% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_ts_design.R
\name{validate_ts_parameters}
\alias{validate_ts_parameters}
\title{Perevirka ta pidhotovka parametriv chasovoho riadu}
\usage{
validate_ts_parameters(x, order, model_type, include.mean)
}
\arguments{
\item{x}{Dani chasovoho riadu}

\item{order}{Spetsyfikatsiia poriadku modeli}

\item{model_type}{Typ modeli (ar, ma, arma, abo arima)}

\item{include.mean}{Chy vkliuchaty serednie/perekhoplennia}
}
\value{
Spysok perevirenykh parametriv ta informatsii pro model
}
\description{
Perevirka ta pidhotovka parametriv chasovoho riadu
}
\keyword{internal}
