% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_main.R
\name{vcov,PMM2fit-method}
\alias{vcov,PMM2fit-method}
\title{Variance-covariance matrix for PMM2fit object}
\usage{
\S4method{vcov}{PMM2fit}(object, ...)
}
\arguments{
\item{object}{PMM2fit object}

\item{...}{Additional arguments (not used)}
}
\value{
Numeric matrix of the same dimension as the number of coefficients
}
\description{
Returns the asymptotic covariance matrix \eqn{\sigma^2 g_2 (X^\top X)^{-1}},
where \eqn{g_2 = 1 - c_3^2 / (2 + c_4)} is the PMM2 variance reduction factor.
Under Gaussian errors \eqn{g_2 = 1} and the result equals the OLS covariance matrix.
}
