% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm_main.R
\name{AIC,PMM2fit-method}
\alias{AIC,PMM2fit-method}
\title{Розрахувати AIC для об'єкта PMM2fit}
\usage{
\S4method{AIC}{PMM2fit}(object, ..., k = 2)
}
\arguments{
\item{object}{Об'єкт PMM2fit}

\item{...}{Додаткові аргументи (не використовуються)}

\item{k}{Штраф за параметр, що буде використовуватись; стандартно 2}
}
\value{
Значення AIC
}
\description{
Розрахувати AIC для об'єкта PMM2fit
}
