% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_main.R
\name{AIC,PMM2fit-method}
\alias{AIC,PMM2fit-method}
\title{Rozrakhuvaty AIC dlia ob'iekta PMM2fit}
\usage{
\S4method{AIC}{PMM2fit}(object, ..., k = 2)
}
\arguments{
\item{object}{Ob'iekt PMM2fit}

\item{...}{Dodatkovi arhumenty (ne vykorystovuiutsia)}

\item{k}{Shtraf za parametr, shcho bude vykorystovuvatys; standartno 2}
}
\value{
Znachennia AIC
}
\description{
Rozrakhuvaty AIC dlia ob'iekta PMM2fit
}
