% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_ts_main.R
\name{arma_pmm2}
\alias{arma_pmm2}
\title{Pidihnaty ARMA model za dopomohoiu PMM2 (obhortka)}
\usage{
arma_pmm2(
  x,
  order = c(1, 1),
  method = "pmm2",
  max_iter = 50,
  tol = 1e-06,
  include.mean = TRUE,
  initial = NULL,
  na.action = na.fail,
  regularize = TRUE,
  reg_lambda = 1e-08,
  verbose = FALSE
)
}
\arguments{
\item{x}{Chyslovyi vektor danykh chasovoho riadu}

\item{order}{Spetsyfikatsiia poriadku modeli:
- Dlia AR modelei: odne tsile chyslo (poriadok AR)
- Dlia MA modelei: odne tsile chyslo (poriadok MA)
- Dlia ARMA modelei: vektor c(p, q) (poriadky AR ta MA)
- Dlia ARIMA modelei: vektor c(p, d, q) (poriadky AR, dyferentsiiuvannia ta MA)}

\item{method}{Riadok: metod otsiniuvannia, odyn z "pmm2" (za zamovchuvanniam), "css", "ml", "yw", "ols"}

\item{max_iter}{Tsile chyslo: maksymalna kilkist iteratsii dlia alhorytmu}

\item{tol}{Chyslove: dopusk dlia zbizhnosti}

\item{include.mean}{Lohichne: chy vkliuchaty chlen serednoho (perekhoplennia)}

\item{initial}{Spysok abo vektor pochatkovykh otsinok parametriv (optsionalno)}

\item{na.action}{Funktsiia dlia obrobky vidsutnikh znachen, za zamovchuvanniam - na.fail}

\item{regularize}{Lohichne, dodavaty mali znachennia do diahonali dlia chyslovoi stabilnosti}

\item{reg_lambda}{Parametr rehuliaryzatsii (iakshcho regularize=TRUE)}

\item{verbose}{Lohichne: chy vyvodyty informatsiiu pro prohres}
}
\description{
Pidihnaty ARMA model za dopomohoiu PMM2 (obhortka)
}
