% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_classes.R
\docType{class}
\name{BasePMM2-class}
\alias{BasePMM2-class}
\title{Base S4 class for storing PMM2 model results}
\description{
Base S4 class for storing PMM2 model results
}
\section{Slots}{

\describe{
\item{\code{coefficients}}{numeric vector of estimated parameters}

\item{\code{residuals}}{numeric vector of final residuals}

\item{\code{m2}}{numeric second central moment of initial residuals}

\item{\code{m3}}{numeric third central moment of initial residuals}

\item{\code{m4}}{numeric fourth central moment of initial residuals}

\item{\code{convergence}}{logical or integer code indicating whether algorithm converged}

\item{\code{iterations}}{numeric number of iterations performed}

\item{\code{call}}{original function call}
}}

