% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_classes.R
\docType{class}
\name{BasePMM2-class}
\alias{BasePMM2-class}
\title{Bazovyi klas S4 dlia zberihannia rezultativ PMM2 modelei}
\description{
Bazovyi klas S4 dlia zberihannia rezultativ PMM2 modelei
}
\section{Slots}{

\describe{
\item{\code{coefficients}}{chyslovyi vektor otsinenykh parametriv}

\item{\code{residuals}}{chyslovyi vektor kintsevykh zalyshkiv}

\item{\code{m2}}{chyslovyi druhyi tsentralnyi moment pochatkovykh zalyshkiv}

\item{\code{m3}}{chyslovyi tretii tsentralnyi moment pochatkovykh zalyshkiv}

\item{\code{m4}}{chyslovyi chetvertyi tsentralnyi moment pochatkovykh zalyshkiv}

\item{\code{convergence}}{lohichnyi abo tsilyi kod, shcho vkazuie, chy alhorytm zbihsia}

\item{\code{iterations}}{chyslova kilkist vykonanykh iteratsii}

\item{\code{call}}{oryhinalnyi vyklyk funktsii}
}}

