% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_classes.R
\docType{class}
\name{BasePMM2-class}
\alias{BasePMM2-class}
\title{Базовий клас S4 для зберігання результатів PMM2 моделей}
\description{
Базовий клас S4 для зберігання результатів PMM2 моделей
}
\section{Slots}{

\describe{
\item{\code{coefficients}}{числовий вектор оцінених параметрів}

\item{\code{residuals}}{числовий вектор кінцевих залишків}

\item{\code{m2}}{числовий другий центральний момент початкових залишків}

\item{\code{m3}}{числовий третій центральний момент початкових залишків}

\item{\code{m4}}{числовий четвертий центральний момент початкових залишків}

\item{\code{convergence}}{логічний або цілий код, що вказує, чи алгоритм збігся}

\item{\code{iterations}}{числова кількість виконаних ітерацій}

\item{\code{call}}{оригінальний виклик функції}
}}

