% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_main.R
\name{coef,PMM2fit-method}
\alias{coef,PMM2fit-method}
\title{Extract coefficients from PMM2fit object}
\usage{
\S4method{coef}{PMM2fit}(object, ...)
}
\arguments{
\item{object}{PMM2fit object}

\item{...}{Additional arguments (not used)}
}
\value{
Vector of coefficients
}
\description{
Extract coefficients from PMM2fit object
}
