% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_main.R
\name{coef,PMM2fit-method}
\alias{coef,PMM2fit-method}
\title{Vytiahnuty koefitsiienty z ob'iekta PMM2fit}
\usage{
\S4method{coef}{PMM2fit}(object, ...)
}
\arguments{
\item{object}{Ob'iekt PMM2fit}

\item{...}{Dodatkovi arhumenty (ne vykorystovuiutsia)}
}
\value{
Vektor koefitsiientiv
}
\description{
Vytiahnuty koefitsiienty z ob'iekta PMM2fit
}
