% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_main.R
\name{coef,PMM2fit-method}
\alias{coef,PMM2fit-method}
\title{Витягнути коефіцієнти з об'єкта PMM2fit}
\usage{
\S4method{coef}{PMM2fit}(object, ...)
}
\arguments{
\item{object}{Об'єкт PMM2fit}

\item{...}{Додаткові аргументи (не використовуються)}
}
\value{
Вектор коефіцієнтів
}
\description{
Витягнути коефіцієнти з об'єкта PMM2fit
}
