% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_ts_methods.R
\name{coef,TS2fit-method}
\alias{coef,TS2fit-method}
\title{Extract coefficients from TS2fit object}
\usage{
\S4method{coef}{TS2fit}(object, ...)
}
\arguments{
\item{object}{TS2fit object}

\item{...}{Additional arguments (not used)}
}
\value{
Named vector of coefficients
}
\description{
Extract coefficients from TS2fit object
}
