% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm_ts_methods.R
\name{coef,TS2fit-method}
\alias{coef,TS2fit-method}
\title{Extract coefficients from a TS2fit object}
\usage{
\S4method{coef}{TS2fit}(object, ...)
}
\arguments{
\item{object}{A TS2fit object}

\item{...}{Additional arguments (not used)}
}
\value{
A named vector of coefficients
}
\description{
Extract coefficients from a TS2fit object
}
