% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_ts_methods.R
\name{coef,TS2fit-method}
\alias{coef,TS2fit-method}
\title{Vytiahnuty koefitsiienty z ob'iekta TS2fit}
\usage{
\S4method{coef}{TS2fit}(object, ...)
}
\arguments{
\item{object}{Ob'iekt TS2fit}

\item{...}{Dodatkovi arhumenty (ne vykorystovuiutsia)}
}
\value{
Imenovanyi vektor koefitsiientiv
}
\description{
Vytiahnuty koefitsiienty z ob'iekta TS2fit
}
