% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_ts_methods.R
\name{compare_arima_methods}
\alias{compare_arima_methods}
\title{Compare ARIMA methods}
\usage{
compare_arima_methods(
  x,
  order = c(1, 1, 1),
  include.mean = TRUE,
  pmm2_args = list()
)
}
\arguments{
\item{x}{Numeric vector of time series data}

\item{order}{Model order specification (see ts_pmm2 for format)}

\item{include.mean}{Logical, whether to include intercept term}

\item{pmm2_args}{List of additional arguments to pass to ts_pmm2()}
}
\description{
Compare ARIMA methods
}
