% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_ts_methods.R
\name{compare_arma_methods}
\alias{compare_arma_methods}
\title{Порівняти методи ARMA}
\usage{
compare_arma_methods(
  x,
  order = c(1, 1),
  include.mean = TRUE,
  pmm2_args = list()
)
}
\arguments{
\item{x}{Числовий вектор даних часового ряду}

\item{order}{Специфікація порядку моделі (див. ts_pmm2 для формату)}

\item{include.mean}{Логічне, чи включати член перехоплення}

\item{pmm2_args}{Список додаткових аргументів для передачі в ts_pmm2()}
}
\description{
Порівняти методи ARMA
}
