% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_ts_methods.R
\name{compare_arma_methods}
\alias{compare_arma_methods}
\title{Porivniaty metody ARMA}
\usage{
compare_arma_methods(
  x,
  order = c(1, 1),
  include.mean = TRUE,
  pmm2_args = list()
)
}
\arguments{
\item{x}{Chyslovyi vektor danykh chasovoho riadu}

\item{order}{Spetsyfikatsiia poriadku modeli (dyv. ts_pmm2 dlia formatu)}

\item{include.mean}{Lohichne, chy vkliuchaty chlen perekhoplennia}

\item{pmm2_args}{Spysok dodatkovykh arhumentiv dlia peredachi v ts_pmm2()}
}
\description{
Porivniaty metody ARMA
}
