% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm_ts_methods.R
\name{compare_ma_methods}
\alias{compare_ma_methods}
\title{Compare MA models}
\usage{
compare_ma_methods(x, order = 1, include.mean = TRUE, pmm2_args = list())
}
\arguments{
\item{x}{A numeric vector of time series data}

\item{order}{Model order specification (see ts_pmm2 for format)}

\item{include.mean}{Logical, whether to include an intercept term}

\item{pmm2_args}{List of additional arguments to pass to ts_pmm2()}
}
\description{
Compare MA models
}
