% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm_ts_methods.R
\name{compare_ma_methods}
\alias{compare_ma_methods}
\title{Порівняти методи MA}
\usage{
compare_ma_methods(x, order = 1, include.mean = TRUE, pmm2_args = list())
}
\arguments{
\item{x}{Числовий вектор даних часового ряду}

\item{order}{Специфікація порядку моделі (див. ts_pmm2 для формату)}

\item{include.mean}{Логічне, чи включати член перехоплення}

\item{pmm2_args}{Список додаткових аргументів для передачі в ts_pmm2()}
}
\description{
Порівняти методи MA
}
