% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_ts_methods.R
\name{compare_ts_methods}
\alias{compare_ts_methods}
\title{Porivniaty PMM2 z klasychnymy metodamy otsiniuvannia chasovykh riadiv}
\usage{
compare_ts_methods(
  x,
  order,
  model_type = c("ar", "ma", "arma", "arima"),
  include.mean = TRUE,
  pmm2_args = list()
)
}
\arguments{
\item{x}{Chyslovyi vektor danykh chasovoho riadu}

\item{order}{Spetsyfikatsiia poriadku modeli (dyv. ts_pmm2 dlia formatu)}

\item{model_type}{Typ modeli: "ar", "ma", "arma", abo "arima"}

\item{include.mean}{Lohichne, chy vkliuchaty chlen perekhoplennia}

\item{pmm2_args}{Spysok dodatkovykh arhumentiv dlia peredachi v ts_pmm2()}
}
\value{
Spysok z pidihnanymy modeliamy ta tablytsiamy porivniannia
}
\description{
Porivniaty PMM2 z klasychnymy metodamy otsiniuvannia chasovykh riadiv
}
