% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_main.R
\name{compare_with_ols}
\alias{compare_with_ols}
\title{Compare PMM2 with OLS}
\usage{
compare_with_ols(formula, data, pmm2_args = list())
}
\arguments{
\item{formula}{Model formula}

\item{data}{Data frame}

\item{pmm2_args}{List of arguments to pass to lm_pmm2()}
}
\value{
List with OLS and PMM2 fit objects
}
\description{
Compare PMM2 with OLS
}
