% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm_main.R
\name{compare_with_ols}
\alias{compare_with_ols}
\title{Порівняти PMM2 з OLS}
\usage{
compare_with_ols(formula, data, pmm2_args = list())
}
\arguments{
\item{formula}{Формула моделі}

\item{data}{Фрейм даних}

\item{pmm2_args}{Список аргументів для передачі в lm_pmm2()}
}
\value{
Список з об'єктами підгонки OLS та PMM2
}
\description{
Порівняти PMM2 з OLS
}
