% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_ts_design.R
\name{compute_ts_residuals}
\alias{compute_ts_residuals}
\title{Обчислити кінцеві залишки для моделей часових рядів}
\usage{
compute_ts_residuals(coefs, model_info)
}
\arguments{
\item{coefs}{Оцінені коефіцієнти}

\item{model_info}{Інформація про модель}
}
\value{
Вектор залишків
}
\description{
Обчислити кінцеві залишки для моделей часових рядів
}
\keyword{internal}
