% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_ts_design.R
\name{compute_ts_residuals}
\alias{compute_ts_residuals}
\title{Obchyslyty kintsevi zalyshky dlia modelei chasovykh riadiv}
\usage{
compute_ts_residuals(coefs, model_info)
}
\arguments{
\item{coefs}{Otsineni koefitsiienty}

\item{model_info}{Informatsiia pro model}
}
\value{
Vektor zalyshkiv
}
\description{
Obchyslyty kintsevi zalyshky dlia modelei chasovykh riadiv
}
\keyword{internal}
