% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_ts_design.R
\name{create_ar_matrix}
\alias{create_ar_matrix}
\title{Create design matrix for AR model}
\usage{
create_ar_matrix(x, p)
}
\arguments{
\item{x}{Centered time series}

\item{p}{AR order}
}
\value{
Design matrix with lagged values
}
\description{
Create design matrix for AR model
}
\keyword{internal}
