% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_ts_design.R
\name{create_ar_matrix}
\alias{create_ar_matrix}
\title{Створення матриці дизайну для AR моделі}
\usage{
create_ar_matrix(x, p)
}
\arguments{
\item{x}{центрований часовий ряд}

\item{p}{порядок AR}
}
\value{
Матриця дизайну з лаговими значеннями
}
\description{
Створення матриці дизайну для AR моделі
}
\keyword{internal}
