% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_ts_design.R
\name{create_ar_matrix}
\alias{create_ar_matrix}
\title{Stvorennia matrytsi dyzainu dlia AR modeli}
\usage{
create_ar_matrix(x, p)
}
\arguments{
\item{x}{tsentrovanyi chasovyi riad}

\item{p}{poriadok AR}
}
\value{
Matrytsia dyzainu z lahovymy znachenniamy
}
\description{
Stvorennia matrytsi dyzainu dlia AR modeli
}
\keyword{internal}
