% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_ts_design.R
\name{create_ts_design_matrix}
\alias{create_ts_design_matrix}
\title{Create design matrix for time series}
\usage{
create_ts_design_matrix(x, model_info, innovations = NULL)
}
\arguments{
\item{x}{Time series data}

\item{model_info}{List with model parameters}

\item{innovations}{Optional innovations/residuals for MA components}
}
\value{
List with design matrix, response variable, and other components
}
\description{
Create design matrix for time series
}
\keyword{internal}
