% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_ts_design.R
\name{create_ts_design_matrix}
\alias{create_ts_design_matrix}
\title{Stvorennia dyzain-matrytsi dlia chasovykh riadiv}
\usage{
create_ts_design_matrix(x, model_info, innovations = NULL)
}
\arguments{
\item{x}{Dani chasovoho riadu}

\item{model_info}{Spysok z parametramy modeli}

\item{innovations}{Optsionalni innovatsii/zalyshky dlia MA komponentiv}
}
\value{
Spysok z dyzain-matrytseiu, zminnoiu vidhuku ta inshymy komponentamy
}
\description{
Stvorennia dyzain-matrytsi dlia chasovykh riadiv
}
\keyword{internal}
