% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_package.R
\docType{package}
\name{EstemPMM-package}
\alias{EstemPMM}
\alias{EstemPMM-package}
\title{PMM2: Polinomialnyi metod maksymizatsii dlia robastnoi rehresii ta chasovykh riadiv}
\description{
Paket PMM2 proponuie robastni metody dlia otsiniuvannia parametriv liniinykh
modelei ta modelei chasovykh riadiv, robastnykh do nehausivskykh pomylok.
}
\section{Funktsii liniinoi rehresii}{


\code{\link{lm_pmm2}} - Pidhonka liniinoi modeli za dopomohoiu PMM2

\code{\link{compare_with_ols}} - Porivniannia PMM2 z OLS
}

\section{Funktsii chasovykh riadiv}{


\code{\link{ts_pmm2}} - Zahalna funktsiia dlia pidhonky modelei chasovykh riadiv za dopomohoiu PMM2

\code{\link{ar_pmm2}} - Pidhonka AR modelei

\code{\link{ma_pmm2}} - Pidhonka MA modelei

\code{\link{arma_pmm2}} - Pidhonka ARMA modelei

\code{\link{arima_pmm2}} - Pidhonka ARIMA modelei

\code{\link{compare_ts_methods}} - Porivniannia PMM2 z klasychnymy metodamy
}

\section{Statystychnyi vysnovok}{


\code{\link{pmm2_inference}} - Butstrep-vysnovok dlia liniinykh modelei

\code{\link{ts_pmm2_inference}} - Butstrep-vysnovok dlia modelei chasovykh riadiv
}

\section{Utylity}{


\code{\link{pmm_skewness}} - Obchyslennia asymetrii

\code{\link{pmm_kurtosis}} - Obchyslennia ekstsesu

\code{\link{compute_moments}} - Obchyslennia momentiv ta kumuliantiv
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SZabolotnii/EstemPMM}
  \item Report bugs at \url{https://github.com/SZabolotnii/EstemPMM/issues}
}

}
\author{
\strong{Maintainer}: Serhii Zabolotnii \email{zabolotniua@gmail.com} (\href{https://orcid.org/0000-0003-0242-2234}{ORCID})

}
\keyword{internal}
