% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_main.R
\name{fitted,PMM2fit-method}
\alias{fitted,PMM2fit-method}
\title{Extract fitted values from PMM2fit object}
\usage{
\S4method{fitted}{PMM2fit}(object, data = NULL, ...)
}
\arguments{
\item{object}{PMM2fit object}

\item{data}{Optional data source for model reconstruction, if object does not contain saved data}

\item{...}{Additional arguments (not used)}
}
\value{
Vector of fitted values
}
\description{
Extract fitted values from PMM2fit object
}
