% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm_main.R
\name{fitted,PMM2fit-method}
\alias{fitted,PMM2fit-method}
\title{Extract fitted values from a PMM2fit object}
\usage{
\S4method{fitted}{PMM2fit}(object, ...)
}
\arguments{
\item{object}{A PMM2fit object}

\item{...}{Additional arguments (not used)}
}
\value{
A vector of fitted values
}
\description{
Extract fitted values from a PMM2fit object
}
