% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm_main.R
\name{fitted,PMM2fit-method}
\alias{fitted,PMM2fit-method}
\title{Витягнути підігнані значення з об'єкта PMM2fit}
\usage{
\S4method{fitted}{PMM2fit}(object, data = NULL, ...)
}
\arguments{
\item{object}{Об'єкт PMM2fit}

\item{...}{Додаткові аргументи (не використовуються)}
}
\value{
Вектор підігнаних значень
}
\description{
Витягнути підігнані значення з об'єкта PMM2fit
}
