% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_ts_methods.R
\name{fitted,TS2fit-method}
\alias{fitted,TS2fit-method}
\title{Extract fitted values from TS2fit object}
\usage{
\S4method{fitted}{TS2fit}(object, ...)
}
\arguments{
\item{object}{TS2fit object}

\item{...}{Additional arguments (not used)}
}
\value{
Vector of fitted values
}
\description{
Extract fitted values from TS2fit object
}
