% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_ts_methods.R
\name{fitted,TS2fit-method}
\alias{fitted,TS2fit-method}
\title{Витягнути підігнані значення з об'єкта TS2fit}
\usage{
\S4method{fitted}{TS2fit}(object, ...)
}
\arguments{
\item{object}{Об'єкт TS2fit}

\item{...}{Додаткові аргументи (не використовуються)}
}
\value{
Вектор підігнаних значень
}
\description{
Витягнути підігнані значення з об'єкта TS2fit
}
