% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_main.R
\name{fitted_values}
\alias{fitted_values}
\title{Helper function for extracting fitted values}
\usage{
fitted_values(object, data = NULL)
}
\arguments{
\item{object}{PMM2fit object}
}
\value{
Vector of fitted values
}
\description{
Helper function for extracting fitted values
}
\keyword{internal}
