% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm_classes.R
\name{fitted_values}
\alias{fitted_values}
\title{Helper function to extract fitted values}
\usage{
fitted_values(object)
}
\arguments{
\item{object}{A PMM2fit object}
}
\value{
A vector of fitted values
}
\description{
Helper function to extract fitted values
}
\keyword{internal}
