% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_main.R
\name{fitted_values}
\alias{fitted_values}
\title{Допоміжна функція для витягнення підігнаних значень}
\usage{
fitted_values(object, data = NULL)
}
\arguments{
\item{object}{Об'єкт PMM2fit}
}
\value{
Вектор підігнаних значень
}
\description{
Допоміжна функція для витягнення підігнаних значень
}
\keyword{internal}
