% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_ts_methods.R
\name{get_ar_fitted}
\alias{get_ar_fitted}
\title{Get fitted values for AR model}
\usage{
get_ar_fitted(object)
}
\arguments{
\item{object}{TS2fit object with model_type="ar"}
}
\value{
Vector of fitted values
}
\description{
Get fitted values for AR model
}
\keyword{internal}
