% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_ts_methods.R
\name{get_ar_fitted}
\alias{get_ar_fitted}
\title{Otrymaty pidihnani znachennia dlia AR modeli}
\usage{
get_ar_fitted(object)
}
\arguments{
\item{object}{Ob'iekt TS2fit z model_type="ar"}
}
\value{
Vektor pidihnanykh znachen
}
\description{
Otrymaty pidihnani znachennia dlia AR modeli
}
\keyword{internal}
