% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm_ts_methods.R
\name{get_ar_fitted}
\alias{get_ar_fitted}
\title{Отримати підігнані значення для AR моделі}
\usage{
get_ar_fitted(object)
}
\arguments{
\item{object}{Об'єкт TS2fit з model_type="ar"}
}
\value{
Вектор підігнаних значень
}
\description{
Отримати підігнані значення для AR моделі
}
\keyword{internal}
