% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm_ts.R
\name{get_initial_estimates}
\alias{get_initial_estimates}
\title{Get initial parameter estimates for time series models}
\usage{
get_initial_estimates(
  model_params,
  initial = NULL,
  method = "pmm2",
  verbose = FALSE
)
}
\arguments{
\item{model_params}{Validated model parameters from validate_ts_parameters}

\item{initial}{Optional user-provided initial estimates}

\item{method}{Estimation method}

\item{verbose}{Print verbose output}
}
\value{
List containing initial parameter estimates and innovations
}
\description{
Get initial parameter estimates for time series models
}
\keyword{internal}
