% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_ts_design.R
\name{get_yw_estimates}
\alias{get_yw_estimates}
\title{Otrymaty otsinky Yula-Volkera dlia AR(p)}
\usage{
get_yw_estimates(x, p)
}
\arguments{
\item{x}{chyslovyi vektor}

\item{p}{tsile znachennia poriadku AR}
}
\value{
chyslovyi vektor dovzhyny p (koefitsiienty AR)
}
\description{
Otrymaty otsinky Yula-Volkera dlia AR(p)
}
\keyword{internal}
