% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm_utils.R
\name{get_yw_estimates}
\alias{get_yw_estimates}
\title{Get Yule-Walker estimates for AR model}
\usage{
get_yw_estimates(x, p)
}
\arguments{
\item{x}{centered time series}

\item{p}{AR order}
}
\value{
Vector of AR coefficients
}
\description{
Get Yule-Walker estimates for AR model
}
\keyword{internal}
