% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm_ts_design.R
\name{get_yw_estimates}
\alias{get_yw_estimates}
\title{Отримати оцінки Юла-Волкера для AR(p)}
\usage{
get_yw_estimates(x, p)
}
\arguments{
\item{x}{числовий вектор}

\item{p}{ціле значення порядку AR}
}
\value{
числовий вектор довжини p (коефіцієнти AR)
}
\description{
Отримати оцінки Юла-Волкера для AR(p)
}
\keyword{internal}
