% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm_main.R
\name{lm_pmm2}
\alias{lm_pmm2}
\title{pmm2: Master function for PMM2 (S=2)}
\usage{
lm_pmm2(formula, data, max_iter = 50, tol = 1e-06, regularize = TRUE)
}
\arguments{
\item{formula}{R formula}

\item{data}{data.frame}

\item{max_iter}{integer: iteration limit}

\item{tol}{numeric: tolerance for convergence}

\item{regularize}{logical: add small value to diagonal for numerical stability}
}
\value{
An S4 \code{PMM2fit} object
}
\description{
pmm2: Master function for PMM2 (S=2)
}
\details{
\enumerate{
\item Fits OLS => b_ols
\item computes m2,m3,m4 from OLS residuals
\item calls internal iteration approach
}
}
