% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_main.R
\name{lm_pmm2}
\alias{lm_pmm2}
\title{pmm2: Holovna funktsiia dlia PMM2 (S=2)}
\usage{
lm_pmm2(
  formula,
  data,
  max_iter = 50,
  tol = 1e-06,
  regularize = TRUE,
  reg_lambda = 1e-08,
  na.action = na.fail,
  weights = NULL,
  verbose = FALSE
)
}
\arguments{
\item{formula}{Formula R dlia modeli}

\item{data}{data.frame, shcho mistyt zminni u formuli}

\item{max_iter}{tsile: maksymalna kilkist iteratsii dlia alhorytmu}

\item{tol}{chyslove: dopusk dlia zbizhnosti}

\item{regularize}{lohichne: dodaty male znachennia do diahonali dlia chyslovoi stabilnosti}

\item{reg_lambda}{chyslove: parametr rehuliaryzatsii (iakshcho regularize=TRUE)}

\item{na.action}{funktsiia dlia obrobky vidsutnikh znachen, za zamovchuvanniam - na.fail}

\item{weights}{optsionalnyi vektor vah (poky ne realizovano)}

\item{verbose}{lohichne: chy vyvodyty informatsiiu pro prohres}
}
\value{
Ob'iekt S4 \code{PMM2fit}
}
\description{
Pidhaniaie liniinu model za dopomohoiu metodu maksymizatsii polinomiv (poriadok 2),
iakyi ie robastnym shchodo nehausivskykh pomylok.
}
\details{
Alhorytm PMM2 pratsiuie nastupnym chynom:
\enumerate{
\item Pidhaniaie zvychainu rehresiiu naimenshykh kvadrativ (OLS) dlia otrymannia pochatkovykh otsinok
\item Obchysliuie tsentralni momenty (m2, m3, m4) iz zalyshkiv OLS
\item Iteratyvno pokrashchuie otsinky parametriv za dopomohoiu pidkhodu na osnovi hradiienta
}

PMM2 osoblyvo korysnyi, koly termy pomylok ne ie hausivskymy.
}
\examples{
\dontrun{
# Heneruvaty dani vybirky z t-rozpodilenymy pomylkamy
n <- 100
x <- rnorm(n)
y <- 2 + 3*x + rt(n, df=3)
dat <- data.frame(y=y, x=x)

# Pidihnaty model za dopomohoiu PMM2
fit <- lm_pmm2(y ~ x, data=dat)

# Reziume ta statystychnyi vysnovok
summary(fit, formula=y~x, data=dat)
}
}
