% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_ts_main.R
\name{ma_pmm2}
\alias{ma_pmm2}
\title{Підігнати MA модель за допомогою PMM2 (обгортка)}
\usage{
ma_pmm2(
  x,
  order = 1,
  method = "pmm2",
  max_iter = 50,
  tol = 1e-06,
  include.mean = TRUE,
  initial = NULL,
  na.action = na.fail,
  regularize = TRUE,
  reg_lambda = 1e-08,
  verbose = FALSE
)
}
\arguments{
\item{x}{Числовий вектор даних часового ряду}

\item{order}{Специфікація порядку моделі:
- Для AR моделей: одне ціле число (порядок AR)
- Для MA моделей: одне ціле число (порядок MA)
- Для ARMA моделей: вектор c(p, q) (порядки AR та MA)
- Для ARIMA моделей: вектор c(p, d, q) (порядки AR, диференціювання та MA)}

\item{method}{Рядок: метод оцінювання, один з "pmm2" (за замовчуванням), "css", "ml", "yw", "ols"}

\item{max_iter}{Ціле число: максимальна кількість ітерацій для алгоритму}

\item{tol}{Числове: допуск для збіжності}

\item{include.mean}{Логічне: чи включати член середнього (перехоплення)}

\item{initial}{Список або вектор початкових оцінок параметрів (опціонально)}

\item{na.action}{Функція для обробки відсутніх значень, за замовчуванням - na.fail}

\item{regularize}{Логічне, додавати малі значення до діагоналі для числової стабільності}

\item{reg_lambda}{Параметр регуляризації (якщо regularize=TRUE)}

\item{verbose}{Логічне: чи виводити інформацію про прогрес}
}
\description{
Підігнати MA модель за допомогою PMM2 (обгортка)
}
