% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_ts_methods.R
\name{plot,TS2fit,missing-method}
\alias{plot,TS2fit,missing-method}
\title{Build diagnostic plots for TS2fit objects}
\usage{
\S4method{plot}{TS2fit,missing}(x, y, which = c(1:4), ...)
}
\arguments{
\item{x}{TS2fit object}

\item{y}{Not used (for S4 method compatibility)}

\item{which}{Integer vector indicating which plots to produce}

\item{...}{additional arguments passed to plot functions}
}
\value{
Invisibly returns x
}
\description{
Build diagnostic plots for TS2fit objects
}
