% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm_ts_methods.R
\name{plot,TS2fit,missing-method}
\alias{plot,TS2fit,missing-method}
\title{Побудувати діагностичні графіки для об'єктів TS2fit}
\usage{
\S4method{plot}{TS2fit,missing}(x, y, which = c(1:4), ...)
}
\arguments{
\item{x}{Об'єкт TS2fit}

\item{y}{Не використовується (для сумісності методу S4)}

\item{which}{Цілочисельний вектор, що вказує, які графіки виробляти}

\item{...}{додаткові аргументи, передані функціям побудови графіків}
}
\value{
Невидимо повертає x
}
\description{
Побудувати діагностичні графіки для об'єктів TS2fit
}
