% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_inference.R
\name{plot_pmm2_bootstrap}
\alias{plot_pmm2_bootstrap}
\title{Pobuduvaty hrafiky butstrep-rozpodiliv dlia pidhonky PMM2}
\usage{
plot_pmm2_bootstrap(object, coefficients = NULL)
}
\arguments{
\item{object}{Rezultat z pmm2_inference}

\item{coefficients}{Yaki koefitsiienty pobuduvaty, za zamovchuvanniam usi}
}
\value{
Nevydymo povertaie informatsiiu pro histohramu
}
\description{
Pobuduvaty hrafiky butstrep-rozpodiliv dlia pidhonky PMM2
}
