% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_common.R
\name{pmm2_algorithm}
\alias{pmm2_algorithm}
\title{Universalnyi alhorytm PMM2 dlia vsikh typiv modelei}
\usage{
pmm2_algorithm(
  b_init,
  X,
  y,
  m2,
  m3,
  m4,
  max_iter = 50,
  tol = 1e-06,
  regularize = TRUE,
  reg_lambda = 1e-08,
  verbose = FALSE,
  poly_terms = NULL
)
}
\arguments{
\item{b_init}{Pochatkovi otsinky parametriv}

\item{X}{Matrytsia dyzainu}

\item{y}{Vektor vidhuku}

\item{m2, m3, m4}{Tsentralni momenty}

\item{max_iter}{Maksymalna kilkist iteratsii}

\item{tol}{Dopusk dlia zbizhnosti}

\item{regularize}{Chy dodavaty rehuliaryzatsiiu}

\item{reg_lambda}{Parametr rehuliaryzatsii}

\item{verbose}{Chy vyvodyty informatsiiu pro prohres}

\item{poly_terms}{Poperedno obchysleni koefitsiienty polinoma (spysok z elementamy \code{A}, \code{B}, \code{C});
dozvoliaie peredavaty vlasni znachennia dlia spetsialnykh stsenariiv, inakshe vony obchysliuiutsia z momentiv}
}
\value{
Spysok z rezultatamy otsiniuvannia
}
\description{
Universalnyi alhorytm PMM2 dlia vsikh typiv modelei
}
\keyword{internal}
