% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_inference.R
\name{pmm2_inference}
\alias{pmm2_inference}
\title{Bootstrap inference for PMM2 fit}
\usage{
pmm2_inference(
  object,
  formula,
  data,
  B = 200,
  seed = NULL,
  parallel = FALSE,
  cores = NULL
)
}
\arguments{
\item{object}{object of class PMM2fit}

\item{formula}{the same formula that was used initially}

\item{data}{data frame that was used initially}

\item{B}{number of bootstrap replications}

\item{seed}{(optional) for reproducibility}

\item{parallel}{logical, whether to use parallel computing}

\item{cores}{number of cores to use for parallel computing, defaults to auto-detect}
}
\value{
data.frame with columns: Estimate, Std.Error, t.value, p.value
}
\description{
Bootstrap inference for PMM2 fit
}
