% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm_inference.R
\name{pmm2_inference}
\alias{pmm2_inference}
\title{Bootstrap inference for a PMM2 fit}
\usage{
pmm2_inference(object, formula, data, B = 200, seed = NULL)
}
\arguments{
\item{object}{an object of class PMM2fit}

\item{formula, data}{the same formula and data used originally}

\item{B}{number of bootstrap replications}

\item{seed}{(optional) for reproducibility}
}
\value{
A data.frame with columns: Estimate, Std.Error, t.value, p.value
}
\description{
Bootstrap inference for a PMM2 fit
}
