% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_inference.R
\name{pmm2_inference}
\alias{pmm2_inference}
\title{Butstrep-vysnovok dlia pidhonky PMM2}
\usage{
pmm2_inference(
  object,
  formula,
  data,
  B = 200,
  seed = NULL,
  parallel = FALSE,
  cores = NULL
)
}
\arguments{
\item{object}{ob'iekt klasu PMM2fit}

\item{formula}{ta sama formula, shcho vykorystovuvalasia spochatku}

\item{data}{freim danykh, shcho vykorystovuvavsia spochatku}

\item{B}{kilkist butstrep-replikatsii}

\item{seed}{(optsionalno) dlia vidtvoriuvanosti}

\item{parallel}{lohichne, chy vykorystovuvaty paralelni obchyslennia}

\item{cores}{kilkist iader dlia vykorystannia pry paralelnykh obchyslenniakh, za zamovchuvanniam - avtovyznachennia}
}
\value{
data.frame z stovptsiamy: Estimate, Std.Error, t.value, p.value
}
\description{
Butstrep-vysnovok dlia pidhonky PMM2
}
