% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_inference.R
\name{pmm2_inference}
\alias{pmm2_inference}
\title{Бутстреп-висновок для підгонки PMM2}
\usage{
pmm2_inference(
  object,
  formula,
  data,
  B = 200,
  seed = NULL,
  parallel = FALSE,
  cores = NULL
)
}
\arguments{
\item{object}{об'єкт класу PMM2fit}

\item{formula}{та сама формула, що використовувалася спочатку}

\item{data}{фрейм даних, що використовувався спочатку}

\item{B}{кількість бутстреп-реплікацій}

\item{seed}{(опціонально) для відтворюваності}

\item{parallel}{логічне, чи використовувати паралельні обчислення}

\item{cores}{кількість ядер для використання при паралельних обчисленнях, за замовчуванням - автовизначення}
}
\value{
data.frame з стовпцями: Estimate, Std.Error, t.value, p.value
}
\description{
Бутстреп-висновок для підгонки PMM2
}
