% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_utils.R
\name{pmm2_variance_factor}
\alias{pmm2_variance_factor}
\title{Calculate theoretical skewness, kurtosis coefficients and variance reduction factor}
\usage{
pmm2_variance_factor(m2, m3, m4)
}
\arguments{
\item{m2, m3, m4}{central moments of second, third and fourth orders}
}
\value{
List with fields \code{c3}, \code{c4} and \code{g}
}
\description{
Calculate theoretical skewness, kurtosis coefficients and variance reduction factor
}
