% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_utils.R
\name{pmm2_variance_factor}
\alias{pmm2_variance_factor}
\title{Obchyslyty teoretychni koefitsiienty asymetrii, ekstsesu ta faktor zmenshennia dyspersii}
\usage{
pmm2_variance_factor(m2, m3, m4)
}
\arguments{
\item{m2, m3, m4}{tsentralni momenty druhoho, tretoho ta chetvertoho poriadkiv}
}
\value{
Spysok z poliamy \code{c3}, \code{c4} ta \code{g}
}
\description{
Obchyslyty teoretychni koefitsiienty asymetrii, ekstsesu ta faktor zmenshennia dyspersii
}
