% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_utils.R
\name{pmm2_variance_matrices}
\alias{pmm2_variance_matrices}
\title{Calculate theoretical variance matrices for OLS and PMM2}
\usage{
pmm2_variance_matrices(X, m2, m3, m4)
}
\arguments{
\item{X}{Design matrix with column of ones}

\item{m2, m3, m4}{central moments of OLS residuals}
}
\value{
List with fields \code{ols}, \code{pmm2}, \code{c3}, \code{c4}, \code{g}
}
\description{
Calculate theoretical variance matrices for OLS and PMM2
}
