% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_classes.R, R/pmm2_package.R
\docType{class}
\name{PMM2fit-class}
\alias{PMM2fit-class}
\title{Klas S4 dlia zberihannia rezultativ PMM2 rehresiinoi modeli}
\description{
Klas dlia zberihannia rezultativ otsiniuvannia liniinykh modelei za dopomohoiu PMM2
}
\section{Slots}{

\describe{
\item{\code{coefficients}}{chyslovyi vektor otsinenykh parametriv}

\item{\code{residuals}}{chyslovyi vektor kintsevykh zalyshkiv}

\item{\code{m2}}{chyslovyi druhyi tsentralnyi moment pochatkovykh zalyshkiv}

\item{\code{m3}}{chyslovyi tretii tsentralnyi moment pochatkovykh zalyshkiv}

\item{\code{m4}}{chyslovyi chetvertyi tsentralnyi moment pochatkovykh zalyshkiv}

\item{\code{convergence}}{lohichnyi abo tsilyi kod, shcho vkazuie, chy alhorytm zbihsia}

\item{\code{iterations}}{chyslova kilkist vykonanykh iteratsii}

\item{\code{call}}{oryhinalnyi vyklyk funktsii}
}}

\section{Sloty}{

\describe{
\item{coefficients}{Otsineni koefitsiienty}
\item{residuals}{Kintsevi zalyshky}
\item{m2}{Druhyi tsentralnyi moment}
\item{m3}{Tretii tsentralnyi moment}
\item{m4}{Chetvertyi tsentralnyi moment}
\item{convergence}{Status zbizhnosti}
\item{iterations}{Kilkist vykonanykh iteratsii}
\item{call}{Oryhinalnyi vyklyk}
}
}

