% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm_classes.R
\docType{class}
\name{PMM2fit-class}
\alias{PMM2fit-class}
\title{An S4 class to store PMM2 fit results}
\description{
An S4 class to store PMM2 fit results
}
\section{Slots}{

\describe{
\item{\code{coefficients}}{numeric vector of fitted parameters}

\item{\code{residuals}}{numeric vector of final residuals}

\item{\code{m2,m3,m4}}{numeric central moments from the initial residuals}

\item{\code{convergence}}{logical or integer code}
}}

