% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_main.R
\name{predict,PMM2fit-method}
\alias{predict,PMM2fit-method}
\title{Prediction method for PMM2fit objects}
\usage{
\S4method{predict}{PMM2fit}(object, newdata = NULL, debug = FALSE, ...)
}
\arguments{
\item{object}{PMM2fit object}

\item{newdata}{New data frame for prediction}

\item{debug}{Logical value, whether to output debug information}

\item{...}{additional arguments (not used)}
}
\value{
Vector of predictions
}
\description{
Prediction method for PMM2fit objects
}
