% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_main.R
\name{predict,PMM2fit-method}
\alias{predict,PMM2fit-method}
\title{Метод прогнозування для об'єктів PMM2fit}
\usage{
\S4method{predict}{PMM2fit}(object, newdata = NULL, debug = FALSE, ...)
}
\arguments{
\item{object}{Об'єкт PMM2fit}

\item{newdata}{Новий фрейм даних для прогнозування}

\item{debug}{Логічне значення, чи виводити дебаг-інформацію}

\item{...}{додаткові аргументи (не використовуються)}
}
\value{
Вектор прогнозів
}
\description{
Метод прогнозування для об'єктів PMM2fit
}
