% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_main.R
\name{predict,PMM2fit-method}
\alias{predict,PMM2fit-method}
\title{Metod prohnozuvannia dlia ob'iektiv PMM2fit}
\usage{
\S4method{predict}{PMM2fit}(object, newdata = NULL, debug = FALSE, ...)
}
\arguments{
\item{object}{Ob'iekt PMM2fit}

\item{newdata}{Novyi freim danykh dlia prohnozuvannia}

\item{debug}{Lohichne znachennia, chy vyvodyty debah-informatsiiu}

\item{...}{dodatkovi arhumenty (ne vykorystovuiutsia)}
}
\value{
Vektor prohnoziv
}
\description{
Metod prohnozuvannia dlia ob'iektiv PMM2fit
}
