% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm_classes.R
\name{predict,PMM2fit-method}
\alias{predict,PMM2fit-method}
\title{Predict method for PMM2 fit objects}
\usage{
\S4method{predict}{PMM2fit}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{A PMM2fit object}

\item{newdata}{A data frame in which to look for variables with which to predict}

\item{...}{additional arguments (not used)}
}
\value{
A vector of predictions
}
\description{
Predict method for PMM2 fit objects
}
