% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_ts_methods.R
\name{predict,TS2fit-method}
\alias{predict,TS2fit-method}
\title{Метод прогнозування для об'єктів TS2fit}
\usage{
\S4method{predict}{TS2fit}(object, n.ahead = 1, ...)
}
\arguments{
\item{object}{Об'єкт TS2fit}

\item{n.ahead}{Кількість кроків вперед для прогнозування}

\item{...}{додаткові аргументи (не використовуються)}
}
\value{
Вектор або список прогнозів, залежно від типу моделі
}
\description{
Метод прогнозування для об'єктів TS2fit
}
