% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm_ts_methods.R
\name{predict,TS2fit-method}
\alias{predict,TS2fit-method}
\title{Predict method for TS2fit objects}
\usage{
\S4method{predict}{TS2fit}(object, n.ahead = 1, ...)
}
\arguments{
\item{object}{A TS2fit object}

\item{n.ahead}{Number of steps ahead to predict}

\item{...}{additional arguments (not used)}
}
\value{
A vector or list of predictions, depending on model type
}
\description{
Predict method for TS2fit objects
}
