% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_ts_methods.R
\name{predict,TS2fit-method}
\alias{predict,TS2fit-method}
\title{Metod prohnozuvannia dlia ob'iektiv TS2fit}
\usage{
\S4method{predict}{TS2fit}(object, n.ahead = 1, ...)
}
\arguments{
\item{object}{Ob'iekt TS2fit}

\item{n.ahead}{Kilkist krokiv vpered dlia prohnozuvannia}

\item{...}{dodatkovi arhumenty (ne vykorystovuiutsia)}
}
\value{
Vektor abo spysok prohnoziv, zalezhno vid typu modeli
}
\description{
Metod prohnozuvannia dlia ob'iektiv TS2fit
}
