% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_main.R
\name{residuals,PMM2fit-method}
\alias{residuals,PMM2fit-method}
\title{Extract residuals from PMM2fit object}
\usage{
\S4method{residuals}{PMM2fit}(object, ...)
}
\arguments{
\item{object}{PMM2fit object}

\item{...}{Additional arguments (not used)}
}
\value{
Vector of residuals
}
\description{
Extract residuals from PMM2fit object
}
