% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm_main.R
\name{residuals,PMM2fit-method}
\alias{residuals,PMM2fit-method}
\title{Витягнути залишки з об'єкта PMM2fit}
\usage{
\S4method{residuals}{PMM2fit}(object, ...)
}
\arguments{
\item{object}{Об'єкт PMM2fit}

\item{...}{Додаткові аргументи (не використовуються)}
}
\value{
Вектор залишків
}
\description{
Витягнути залишки з об'єкта PMM2fit
}
