% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_ts_methods.R
\name{residuals,TS2fit-method}
\alias{residuals,TS2fit-method}
\title{Vytiahnuty zalyshky z ob'iekta TS2fit}
\usage{
\S4method{residuals}{TS2fit}(object, ...)
}
\arguments{
\item{object}{Ob'iekt TS2fit}

\item{...}{Dodatkovi arhumenty (ne vykorystovuiutsia)}
}
\value{
Vektor zalyshkiv (innovatsii)
}
\description{
Vytiahnuty zalyshky z ob'iekta TS2fit
}
