% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm_ts_methods.R
\name{residuals,TS2fit-method}
\alias{residuals,TS2fit-method}
\title{Витягнути залишки з об'єкта TS2fit}
\usage{
\S4method{residuals}{TS2fit}(object, ...)
}
\arguments{
\item{object}{Об'єкт TS2fit}

\item{...}{Додаткові аргументи (не використовуються)}
}
\value{
Вектор залишків (інновацій)
}
\description{
Витягнути залишки з об'єкта TS2fit
}
