% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm_ts_methods.R
\name{residuals,TS2fit-method}
\alias{residuals,TS2fit-method}
\title{Extract residuals from a TS2fit object}
\usage{
\S4method{residuals}{TS2fit}(object, ...)
}
\arguments{
\item{object}{A TS2fit object}

\item{...}{Additional arguments (not used)}
}
\value{
A vector of residuals (innovations)
}
\description{
Extract residuals from a TS2fit object
}
