% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_ma_estimator.R, R/pmm2_ts_main.R
\name{sma_build_design}
\alias{sma_build_design}
\title{Build design matrix for SMA}
\usage{
sma_build_design(intercept, residuals, x, Q, s)

sma_build_design(intercept, residuals, x, Q, s)
}
\arguments{
\item{intercept}{Model intercept}

\item{residuals}{Initial residuals from CSS fit}

\item{x}{Time series data}

\item{Q}{Seasonal MA order}

\item{s}{Seasonal period}
}
\value{
List with X (design matrix) and y (response)

List with design matrix X and dependent variable y
}
\description{
Build design matrix for SMA

Build design matrix for Seasonal MA model
}
\keyword{internal}
