% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_ma_estimator.R
\name{sma_compute_innovations}
\alias{sma_compute_innovations}
\title{Compute SMA innovations (forward recursion)}
\usage{
sma_compute_innovations(x, Theta, Q, s)
}
\arguments{
\item{x}{Time series (centered)}

\item{Theta}{SMA coefficients}

\item{Q}{SMA order}

\item{s}{Seasonal period}
}
\value{
Vector of innovations
}
\description{
Compute SMA innovations (forward recursion)
}
\keyword{internal}
