% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_ts_main.R
\name{sma_compute_innovations}
\alias{sma_compute_innovations}
\title{Compute innovations for Seasonal MA model}
\usage{
sma_compute_innovations(x, theta, Q, s)
}
\arguments{
\item{x}{Centered time series}

\item{theta}{Seasonal MA coefficients}

\item{Q}{Seasonal MA order}

\item{s}{Seasonal period}
}
\value{
Vector of innovations
}
\description{
Compute innovations for Seasonal MA model
}
\keyword{internal}
