% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_classes.R
\docType{class}
\name{SMAPMM2-class}
\alias{SMAPMM2-class}
\title{S4 class for Seasonal MA PMM2 results}
\description{
This class stores the results of fitting a Seasonal Moving Average (SMA)
model using the Polynomial Maximization Method (PMM2).
}
\details{
The SMA(Q)_s model is expressed as
\deqn{y_t = \mu + \epsilon_t + \Theta_1 \epsilon_{t-s} + \Theta_2 \epsilon_{t-2s} + \dots + \Theta_Q \epsilon_{t-Qs}.}

Where:
\itemize{
\item Q is the seasonal MA order
\item s is the seasonal period
\item epsilon_t are innovations
}
}
\section{Slots}{

\describe{
\item{\code{coefficients}}{Estimated seasonal MA coefficients (Theta_1, Theta_2, ..., Theta_Q)}

\item{\code{innovations}}{Estimated innovations (residuals epsilon_t)}

\item{\code{m2}}{Second central moment (variance) of innovations}

\item{\code{m3}}{Third central moment (skewness indicator) of innovations}

\item{\code{m4}}{Fourth central moment (kurtosis indicator) of innovations}

\item{\code{convergence}}{Logical indicating whether PMM2 algorithm converged}

\item{\code{iterations}}{Number of iterations required for convergence}

\item{\code{call}}{The function call that created this object}

\item{\code{model_type}}{Character string "sma"}

\item{\code{intercept}}{Model intercept (mean)}

\item{\code{original_series}}{Original time series data}

\item{\code{order}}{List with Q (seasonal MA order) and s (seasonal period)}
}}

\seealso{
\code{\link{sma_pmm2}} for fitting SMA models
}
