% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_classes.R
\name{summary,PMM2fit-method}
\alias{summary,PMM2fit-method}
\title{Generic summary method for PMM2fit objects}
\usage{
\S4method{summary}{PMM2fit}(object, formula = NULL, data = NULL, B = 100, ...)
}
\arguments{
\item{object}{object of class "PMM2fit"}

\item{formula}{(optional) formula used for the model}

\item{data}{(optional) data used}

\item{B}{number of bootstrap replications for statistical inference}

\item{...}{additional arguments (not used)}
}
\value{
Prints summary to console; returns object (invisibly).
}
\description{
Generic summary method for PMM2fit objects
}
