% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_classes.R
\name{summary,PMM2fit-method}
\alias{summary,PMM2fit-method}
\title{Узагальнений метод summary для об'єктів PMM2fit}
\usage{
\S4method{summary}{PMM2fit}(object, formula = NULL, data = NULL, B = 100, ...)
}
\arguments{
\item{object}{об'єкт класу "PMM2fit"}

\item{formula}{(опціонально) формула, використана для моделі}

\item{data}{(опціонально) використані дані}

\item{B}{кількість бутстреп-реплікацій для статистичного висновку}

\item{...}{додаткові аргументи (не використовуються)}
}
\value{
Виводить резюме на консоль; повертає об'єкт (невидимо).
}
\description{
Узагальнений метод summary для об'єктів PMM2fit
}
