% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm_classes.R
\name{summary,PMM2fit-method}
\alias{summary,PMM2fit-method}
\title{Summarize PMM2 results}
\usage{
\S4method{summary}{PMM2fit}(object, formula = NULL, data = NULL, B = 100)
}
\arguments{
\item{object}{an object of class "PMM2fit"}

\item{formula}{(optional) the formula used for the model}

\item{data}{(optional) the data used}

\item{B}{number of bootstrap replicates for inference}
}
\value{
Prints a summary to console; returns the object (invisibly).
}
\description{
Summarize PMM2 results
}
