% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm_classes.R
\name{summary,TS2fit-method}
\alias{summary,TS2fit-method}
\title{Узагальнений метод summary для об'єктів TS2fit}
\usage{
\S4method{summary}{TS2fit}(object, ...)
}
\arguments{
\item{object}{об'єкт класу "TS2fit" або підкласу}

\item{...}{додаткові аргументи (не використовуються)}
}
\value{
Виводить резюме на консоль; повертає об'єкт (невидимо).
}
\description{
Узагальнений метод summary для об'єктів TS2fit
}
