% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_classes.R
\name{summary,TS2fit-method}
\alias{summary,TS2fit-method}
\title{Uzahalnenyi metod summary dlia ob'iektiv TS2fit}
\usage{
\S4method{summary}{TS2fit}(object, ...)
}
\arguments{
\item{object}{ob'iekt klasu "TS2fit" abo pidklasu}

\item{...}{dodatkovi arhumenty (ne vykorystovuiutsia)}
}
\value{
Vyvodyt reziume na konsol; povertaie ob'iekt (nevydymo).
}
\description{
Uzahalnenyi metod summary dlia ob'iektiv TS2fit
}
