% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_classes.R, R/pmm2_package.R
\docType{class}
\name{TS2fit-class}
\alias{TS2fit-class}
\title{S4 class for PMM2 time-series fit results}
\description{
Common parent of every PMM2 time-series fit class (AR, MA, ARMA,
ARIMA, and the seasonal SAR/SMA/SARMA/SARIMA subclasses). Inherits
the PMM2-specific central moments from
\code{\linkS4class{BasePMM2}} and the time-series slots
(\code{model_type}, \code{intercept}, \code{original_series},
\code{order}) from \code{\linkS4class{PMMtsfit}}. The diamond
inheritance resolves at \code{\linkS4class{PMMfit}}, which both
branches share as their virtual root.

Base class for storing results of time series model estimation using PMM2
}
\section{Slots}{

\describe{
\item{coefficients}{Estimated coefficients}
\item{residuals}{Final residuals}
\item{m2}{Second central moment}
\item{m3}{Third central moment}
\item{m4}{Fourth central moment}
\item{convergence}{Convergence status}
\item{iterations}{Number of iterations performed}
\item{call}{Original call}
\item{model_type}{Model type}
\item{intercept}{Intercept}
\item{original_series}{Original time series}
\item{order}{Model orders}
}
}

