% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_classes.R, R/pmm2_package.R
\docType{class}
\name{TS2fit-class}
\alias{TS2fit-class}
\title{Bazovyi klas S4 dlia zberihannia rezultativ PMM2 modelei chasovykh riadiv}
\description{
Bazovyi klas dlia zberihannia rezultativ otsiniuvannia modelei chasovykh riadiv za dopomohoiu PMM2
}
\section{Slots}{

\describe{
\item{\code{coefficients}}{chyslovyi vektor otsinenykh parametriv}

\item{\code{residuals}}{chyslovyi vektor kintsevykh zalyshkiv}

\item{\code{m2}}{chyslovyi druhyi tsentralnyi moment pochatkovykh zalyshkiv}

\item{\code{m3}}{chyslovyi tretii tsentralnyi moment pochatkovykh zalyshkiv}

\item{\code{m4}}{chyslovyi chetvertyi tsentralnyi moment pochatkovykh zalyshkiv}

\item{\code{convergence}}{lohichnyi abo tsilyi kod, shcho vkazuie, chy alhorytm zbihsia}

\item{\code{iterations}}{chyslova kilkist vykonanykh iteratsii}

\item{\code{call}}{oryhinalnyi vyklyk funktsii}

\item{\code{model_type}}{symvolnyi riadok, shcho vkazuie typ modeli}

\item{\code{intercept}}{chyslove znachennia perekhoplennia}

\item{\code{original_series}}{chyslovyi vektor oryhinalnoho chasovoho riadu}

\item{\code{order}}{spysok parametriv poriadku}
}}

\section{Sloty}{

\describe{
\item{coefficients}{Otsineni koefitsiienty}
\item{residuals}{Kintsevi zalyshky}
\item{m2}{Druhyi tsentralnyi moment}
\item{m3}{Tretii tsentralnyi moment}
\item{m4}{Chetvertyi tsentralnyi moment}
\item{convergence}{Status zbizhnosti}
\item{iterations}{Kilkist vykonanykh iteratsii}
\item{call}{Oryhinalnyi vyklyk}
\item{model_type}{Typ modeli}
\item{intercept}{Perekhoplennia}
\item{original_series}{Oryhinalnyi chasovyi riad}
\item{order}{Poriadky modeli}
}
}

