% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm_ts.R
\docType{class}
\name{TS2fit-class}
\alias{TS2fit-class}
\title{An S4 class to store time series PMM2 fit results (base class)}
\description{
An S4 class to store time series PMM2 fit results (base class)
}
\section{Slots}{

\describe{
\item{\code{coefficients}}{numeric vector of fitted parameters}

\item{\code{residuals}}{numeric vector of final residuals}

\item{\code{m2}}{numeric second central moment from the initial residuals}

\item{\code{m3}}{numeric third central moment from the initial residuals}

\item{\code{m4}}{numeric fourth central moment from the initial residuals}

\item{\code{convergence}}{logical or integer code indicating if the algorithm converged}

\item{\code{iterations}}{integer number of iterations performed}

\item{\code{call}}{the original function call}

\item{\code{model_type}}{character string indicating the type of model}

\item{\code{intercept}}{numeric intercept value}

\item{\code{original_series}}{numeric original time series data}

\item{\code{order}}{list of order parameters}
}}

