% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_ts_main.R
\name{ts_pmm2}
\alias{ts_pmm2}
\title{Pidihnaty model chasovoho riadu za dopomohoiu metodu PMM2}
\usage{
ts_pmm2(
  x,
  order,
  model_type = c("ar", "ma", "arma", "arima"),
  method = "pmm2",
  max_iter = 50,
  tol = 1e-06,
  include.mean = TRUE,
  initial = NULL,
  na.action = na.fail,
  regularize = TRUE,
  reg_lambda = 1e-08,
  verbose = FALSE
)
}
\arguments{
\item{x}{Chyslovyi vektor danykh chasovoho riadu}

\item{order}{Spetsyfikatsiia poriadku modeli:
- Dlia AR modelei: odne tsile chyslo (poriadok AR)
- Dlia MA modelei: odne tsile chyslo (poriadok MA)
- Dlia ARMA modelei: vektor c(p, q) (poriadky AR ta MA)
- Dlia ARIMA modelei: vektor c(p, d, q) (poriadky AR, dyferentsiiuvannia ta MA)}

\item{model_type}{Riadok, shcho vyznachaie typ modeli: "ar", "ma", "arma", abo "arima"}

\item{method}{Riadok: metod otsiniuvannia, odyn z "pmm2" (za zamovchuvanniam), "css", "ml", "yw", "ols"}

\item{max_iter}{Tsile chyslo: maksymalna kilkist iteratsii dlia alhorytmu}

\item{tol}{Chyslove: dopusk dlia zbizhnosti}

\item{include.mean}{Lohichne: chy vkliuchaty chlen serednoho (perekhoplennia)}

\item{initial}{Spysok abo vektor pochatkovykh otsinok parametriv (optsionalno)}

\item{na.action}{Funktsiia dlia obrobky vidsutnikh znachen, za zamovchuvanniam - na.fail}

\item{regularize}{Lohichne, dodavaty mali znachennia do diahonali dlia chyslovoi stabilnosti}

\item{reg_lambda}{Parametr rehuliaryzatsii (iakshcho regularize=TRUE)}

\item{verbose}{Lohichne: chy vyvodyty informatsiiu pro prohres}
}
\value{
Ob'iekt S4 \code{TS2fit} vidpovidnoho pidklasu
}
\description{
Pidihnaty model chasovoho riadu za dopomohoiu metodu PMM2
}
\details{
Alhorytm PMM2 pratsiuie nastupnym chynom:
\enumerate{
\item Pidhaniaie pochatkovu model za dopomohoiu standartnoho metodu (OLS, Yula-Volkera, CSS abo ML)
\item Obchysliuie tsentralni momenty (m2, m3, m4) z pochatkovykh zalyshkiv/innovatsii
\item Vykorystovuie tsi momenty zi spetsializovanym rozv'iazuvachem (pmm2_algorithm) dlia znakhodzhennia
robastnykh otsinok parametriv
}
}
