% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_inference.R
\name{ts_pmm2_inference}
\alias{ts_pmm2_inference}
\title{Butstrep-vysnovok dlia modelei chasovykh riadiv PMM2}
\usage{
ts_pmm2_inference(
  object,
  x = NULL,
  B = 200,
  seed = NULL,
  block_length = NULL,
  method = c("residual", "block"),
  parallel = FALSE,
  cores = NULL,
  debug = FALSE
)
}
\arguments{
\item{object}{ob'iekt klasu TS2fit}

\item{x}{(optsionalno) oryhinalnyi chasovyi riad; iakshcho NULL, vykorystovuie object@original_series}

\item{B}{kilkist butstrep-replikatsii}

\item{seed}{(optsionalno) dlia vidtvoriuvanosti}

\item{block_length}{dovzhyna bloku dlia blokovoho butstrepu; iakshcho NULL, vykorystovuie evrystychne znachennia}

\item{method}{typ butstrepu: "residual" abo "block"}

\item{parallel}{lohichne, chy vykorystovuvaty paralelni obchyslennia}

\item{cores}{kilkist iader dlia paralelnykh obchyslen}

\item{debug}{lohichne, chy vyvodyty dodatkovu diahnostychnu informatsiiu}
}
\value{
data.frame z stovptsiamy: Estimate, Std.Error, t.value, p.value
}
\description{
Butstrep-vysnovok dlia modelei chasovykh riadiv PMM2
}
