% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm_utils.R
\name{update_ma_innovations}
\alias{update_ma_innovations}
\title{Update innovations for MA model}
\usage{
update_ma_innovations(x, ma_coef)
}
\arguments{
\item{x}{centered time series}

\item{ma_coef}{vector of MA coefficients}
}
\value{
vector of innovations
}
\description{
Update innovations for MA model
}
\keyword{internal}
