% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_ts_design.R
\name{update_ma_innovations}
\alias{update_ma_innovations}
\title{Onovyty innovatsii MA modeli}
\usage{
update_ma_innovations(x, ma_coef)
}
\arguments{
\item{x}{tsentrovanyi chasovyi riad}

\item{ma_coef}{vektor koefitsiientiv MA}
}
\value{
vektor innovatsii
}
\description{
Onovyty innovatsii MA modeli
}
\keyword{internal}
