% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm_ts_design.R
\name{update_ma_innovations}
\alias{update_ma_innovations}
\title{Оновити інновації MA моделі}
\usage{
update_ma_innovations(x, ma_coef)
}
\arguments{
\item{x}{центрований часовий ряд}

\item{ma_coef}{вектор коефіцієнтів MA}
}
\value{
вектор інновацій
}
\description{
Оновити інновації MA моделі
}
\keyword{internal}
