% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm2_ts_design.R
\name{validate_ts_parameters}
\alias{validate_ts_parameters}
\title{Validate and prepare time series parameters}
\usage{
validate_ts_parameters(x, order, model_type, include.mean)
}
\arguments{
\item{x}{Time series data}

\item{order}{Model order specification}

\item{model_type}{Model type (ar, ma, arma, or arima)}

\item{include.mean}{Whether to include mean/intercept}
}
\value{
List of validated parameters and model information
}
\description{
Validate and prepare time series parameters
}
\keyword{internal}
