% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm_ts.R
\name{validate_ts_parameters}
\alias{validate_ts_parameters}
\title{Validates the time series model parameters and prepares model information}
\usage{
validate_ts_parameters(x, order, model_type, include.mean)
}
\arguments{
\item{x}{Time series data}

\item{order}{Model order specification}

\item{model_type}{Type of model (ar, ma, arma, or arima)}

\item{include.mean}{Whether to include mean/intercept}
}
\value{
List of validated parameters and model information
}
\description{
Validates the time series model parameters and prepares model information
}
\keyword{internal}
